refactor: 💡 mocks/handlers/ をリファクタリング
index.tsにhandlersを記載してしまっていたので、handlers.tsへ移動

✅ Closes: #25
