+++ src/mocks/handlers/api/resas/population/composition/perYear/handlers.ts
@@ -1,0 +1,4638 @@
+import { delay, http, HttpResponse } from 'msw';
+
+type PrefCode =
+  | 1
+  | 2
+  | 3
+  | 4
+  | 5
+  | 6
+  | 7
+  | 8
+  | 9
+  | 10
+  | 11
+  | 12
+  | 13
+  | 14
+  | 15
+  | 16
+  | 17
+  | 18
+  | 19
+  | 20
+  | 21
+  | 22
+  | 23
+  | 24
+  | 25
+  | 26
+  | 27
+  | 28
+  | 29
+  | 30
+  | 31
+  | 32
+  | 33
+  | 34
+  | 35
+  | 36
+  | 37
+  | 38
+  | 39
+  | 40
+  | 41
+  | 42
+  | 43
+  | 44
+  | 45
+  | 46
+  | 47;
+
+export const mockPopulationCompositionPerYear = {
+  1: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 5039206 },
+          { year: 1965, value: 5171800 },
+          { year: 1970, value: 5184287 },
+          { year: 1975, value: 5338206 },
+          { year: 1980, value: 5575989 },
+          { year: 1985, value: 5679439 },
+          { year: 1990, value: 5643647 },
+          { year: 1995, value: 5692321 },
+          { year: 2000, value: 5683062 },
+          { year: 2005, value: 5627737 },
+          { year: 2010, value: 5506419 },
+          { year: 2015, value: 5381733 },
+          { year: 2020, value: 5224614 },
+          { year: 2025, value: 5016554 },
+          { year: 2030, value: 4791592 },
+          { year: 2035, value: 4546357 },
+          { year: 2040, value: 4280427 },
+          { year: 2045, value: 4004973 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 1681479, rate: 33.37 },
+          { year: 1965, value: 1462123, rate: 28.27 },
+          { year: 1970, value: 1309487, rate: 25.26 },
+          { year: 1975, value: 1312611, rate: 24.59 },
+          { year: 1980, value: 1298324, rate: 23.28 },
+          { year: 1985, value: 1217959, rate: 21.45 },
+          { year: 1990, value: 1034251, rate: 18.33 },
+          { year: 1995, value: 898673, rate: 15.79 },
+          { year: 2000, value: 792352, rate: 13.94 },
+          { year: 2005, value: 719057, rate: 12.78 },
+          { year: 2010, value: 657312, rate: 11.94 },
+          { year: 2015, value: 608296, rate: 11.3 },
+          { year: 2020, value: 555804, rate: 10.64 },
+          { year: 2025, value: 511677, rate: 10.2 },
+          { year: 2030, value: 465307, rate: 9.71 },
+          { year: 2035, value: 423382, rate: 9.31 },
+          { year: 2040, value: 391086, rate: 9.14 },
+          { year: 2045, value: 360177, rate: 8.99 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 3145664, rate: 62.42 },
+          { year: 1965, value: 3460359, rate: 66.91 },
+          { year: 1970, value: 3575731, rate: 68.97 },
+          { year: 1975, value: 3657884, rate: 68.52 },
+          { year: 1980, value: 3823808, rate: 68.58 },
+          { year: 1985, value: 3910729, rate: 68.86 },
+          { year: 1990, value: 3924717, rate: 69.54 },
+          { year: 1995, value: 3942868, rate: 69.27 },
+          { year: 2000, value: 3832902, rate: 67.44 },
+          { year: 2005, value: 3696064, rate: 65.68 },
+          { year: 2010, value: 3482169, rate: 63.24 },
+          { year: 2015, value: 3190804, rate: 59.29 },
+          { year: 2020, value: 2945727, rate: 56.38 },
+          { year: 2025, value: 2781175, rate: 55.44 },
+          { year: 2030, value: 2594718, rate: 54.15 },
+          { year: 2035, value: 2394230, rate: 52.66 },
+          { year: 2040, value: 2140781, rate: 50.01 },
+          { year: 2045, value: 1931265, rate: 48.22 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 212063, rate: 4.21 },
+          { year: 1965, value: 249318, rate: 4.82 },
+          { year: 1970, value: 299069, rate: 5.77 },
+          { year: 1975, value: 366651, rate: 6.87 },
+          { year: 1980, value: 451727, rate: 8.1 },
+          { year: 1985, value: 549487, rate: 9.68 },
+          { year: 1990, value: 674881, rate: 11.96 },
+          { year: 1995, value: 844927, rate: 14.84 },
+          { year: 2000, value: 1031552, rate: 18.15 },
+          { year: 2005, value: 1205692, rate: 21.42 },
+          { year: 2010, value: 1358068, rate: 24.66 },
+          { year: 2015, value: 1558387, rate: 28.96 },
+          { year: 2020, value: 1664023, rate: 31.85 },
+          { year: 2025, value: 1723702, rate: 34.36 },
+          { year: 2030, value: 1731567, rate: 36.14 },
+          { year: 2035, value: 1728745, rate: 38.02 },
+          { year: 2040, value: 1748560, rate: 40.85 },
+          { year: 2045, value: 1713531, rate: 42.79 },
+        ],
+      },
+    ],
+  },
+  2: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1426606 },
+          { year: 1965, value: 1416591 },
+          { year: 1970, value: 1427520 },
+          { year: 1975, value: 1468646 },
+          { year: 1980, value: 1523907 },
+          { year: 1985, value: 1524448 },
+          { year: 1990, value: 1482873 },
+          { year: 1995, value: 1481663 },
+          { year: 2000, value: 1475728 },
+          { year: 2005, value: 1436657 },
+          { year: 2010, value: 1373339 },
+          { year: 2015, value: 1308265 },
+          { year: 2020, value: 1237984 },
+          { year: 2025, value: 1157332 },
+          { year: 2030, value: 1076393 },
+          { year: 2035, value: 993737 },
+          { year: 2040, value: 908974 },
+          { year: 2045, value: 823610 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 513397, rate: 35.99 },
+          { year: 1965, value: 447068, rate: 31.56 },
+          { year: 1970, value: 396883, rate: 27.8 },
+          { year: 1975, value: 380218, rate: 25.89 },
+          { year: 1980, value: 366454, rate: 24.05 },
+          { year: 1985, value: 338554, rate: 22.21 },
+          { year: 1990, value: 289082, rate: 19.49 },
+          { year: 1995, value: 252414, rate: 17.04 },
+          { year: 2000, value: 223141, rate: 15.12 },
+          { year: 2005, value: 198959, rate: 13.85 },
+          { year: 2010, value: 171842, rate: 12.51 },
+          { year: 2015, value: 148208, rate: 11.33 },
+          { year: 2020, value: 129112, rate: 10.43 },
+          { year: 2025, value: 114024, rate: 9.85 },
+          { year: 2030, value: 100253, rate: 9.31 },
+          { year: 2035, value: 87648, rate: 8.82 },
+          { year: 2040, value: 77258, rate: 8.5 },
+          { year: 2045, value: 67472, rate: 8.19 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 848838, rate: 59.5 },
+          { year: 1965, value: 894521, rate: 63.15 },
+          { year: 1970, value: 940235, rate: 65.86 },
+          { year: 1975, value: 977541, rate: 66.56 },
+          { year: 1980, value: 1022786, rate: 67.12 },
+          { year: 1985, value: 1027329, rate: 67.39 },
+          { year: 1990, value: 1000804, rate: 67.49 },
+          { year: 1995, value: 991311, rate: 66.91 },
+          { year: 2000, value: 964661, rate: 65.37 },
+          { year: 2005, value: 910856, rate: 63.4 },
+          { year: 2010, value: 843587, rate: 61.43 },
+          { year: 2015, value: 757867, rate: 57.93 },
+          { year: 2020, value: 676167, rate: 54.62 },
+          { year: 2025, value: 618505, rate: 53.44 },
+          { year: 2030, value: 555479, rate: 51.61 },
+          { year: 2035, value: 494561, rate: 49.77 },
+          { year: 2040, value: 428573, rate: 47.15 },
+          { year: 2045, value: 370849, rate: 45.03 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 64371, rate: 4.51 },
+          { year: 1965, value: 75002, rate: 5.29 },
+          { year: 1970, value: 90402, rate: 6.33 },
+          { year: 1975, value: 110752, rate: 7.54 },
+          { year: 1980, value: 134516, rate: 8.83 },
+          { year: 1985, value: 158547, rate: 10.4 },
+          { year: 1990, value: 191776, rate: 12.93 },
+          { year: 1995, value: 236745, rate: 15.98 },
+          { year: 2000, value: 287099, rate: 19.45 },
+          { year: 2005, value: 326562, rate: 22.73 },
+          { year: 2010, value: 352768, rate: 25.69 },
+          { year: 2015, value: 390940, rate: 29.88 },
+          { year: 2020, value: 412943, rate: 33.36 },
+          { year: 2025, value: 424803, rate: 36.71 },
+          { year: 2030, value: 420661, rate: 39.08 },
+          { year: 2035, value: 411528, rate: 41.41 },
+          { year: 2040, value: 403143, rate: 44.35 },
+          { year: 2045, value: 385289, rate: 46.78 },
+        ],
+      },
+    ],
+  },
+  3: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1448517 },
+          { year: 1965, value: 1411118 },
+          { year: 1970, value: 1371383 },
+          { year: 1975, value: 1385563 },
+          { year: 1980, value: 1421927 },
+          { year: 1985, value: 1433611 },
+          { year: 1990, value: 1416928 },
+          { year: 1995, value: 1419505 },
+          { year: 2000, value: 1416180 },
+          { year: 2005, value: 1385041 },
+          { year: 2010, value: 1330147 },
+          { year: 2015, value: 1279594 },
+          { year: 2020, value: 1210534 },
+          { year: 2025, value: 1161638 },
+          { year: 2030, value: 1096141 },
+          { year: 2035, value: 1028673 },
+          { year: 2040, value: 957788 },
+          { year: 2045, value: 884518 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 501782, rate: 34.64 },
+          { year: 1965, value: 429521, rate: 30.44 },
+          { year: 1970, value: 362118, rate: 26.41 },
+          { year: 1975, value: 338374, rate: 24.42 },
+          { year: 1980, value: 326014, rate: 22.93 },
+          { year: 1985, value: 307800, rate: 21.47 },
+          { year: 1990, value: 269810, rate: 19.04 },
+          { year: 1995, value: 239010, rate: 16.84 },
+          { year: 2000, value: 212470, rate: 15 },
+          { year: 2005, value: 190578, rate: 13.76 },
+          { year: 2010, value: 168804, rate: 12.69 },
+          { year: 2015, value: 150992, rate: 11.8 },
+          { year: 2020, value: 132447, rate: 10.94 },
+          { year: 2025, value: 123249, rate: 10.61 },
+          { year: 2030, value: 111694, rate: 10.19 },
+          { year: 2035, value: 100559, rate: 9.78 },
+          { year: 2040, value: 90940, rate: 9.49 },
+          { year: 2045, value: 81583, rate: 9.22 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 870492, rate: 60.1 },
+          { year: 1965, value: 895550, rate: 63.46 },
+          { year: 1970, value: 908795, rate: 66.27 },
+          { year: 1975, value: 928674, rate: 67.03 },
+          { year: 1980, value: 952388, rate: 66.98 },
+          { year: 1985, value: 955425, rate: 66.64 },
+          { year: 1990, value: 941052, rate: 66.41 },
+          { year: 1995, value: 925175, rate: 65.18 },
+          { year: 2000, value: 899177, rate: 63.49 },
+          { year: 2005, value: 850253, rate: 61.39 },
+          { year: 2010, value: 795780, rate: 59.83 },
+          { year: 2015, value: 734886, rate: 57.43 },
+          { year: 2020, value: 658816, rate: 54.42 },
+          { year: 2025, value: 625142, rate: 53.82 },
+          { year: 2030, value: 576003, rate: 52.55 },
+          { year: 2035, value: 528635, rate: 51.39 },
+          { year: 2040, value: 472243, rate: 49.31 },
+          { year: 2045, value: 420543, rate: 47.54 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 76243, rate: 5.26 },
+          { year: 1965, value: 86047, rate: 6.1 },
+          { year: 1970, value: 100470, rate: 7.33 },
+          { year: 1975, value: 118397, rate: 8.55 },
+          { year: 1980, value: 143400, rate: 10.08 },
+          { year: 1985, value: 170386, rate: 11.89 },
+          { year: 1990, value: 205737, rate: 14.52 },
+          { year: 1995, value: 255256, rate: 17.98 },
+          { year: 2000, value: 303988, rate: 21.47 },
+          { year: 2005, value: 339957, rate: 24.54 },
+          { year: 2010, value: 360498, rate: 27.1 },
+          { year: 2015, value: 386573, rate: 30.21 },
+          { year: 2020, value: 404359, rate: 33.4 },
+          { year: 2025, value: 413247, rate: 35.57 },
+          { year: 2030, value: 408444, rate: 37.26 },
+          { year: 2035, value: 399479, rate: 38.83 },
+          { year: 2040, value: 394605, rate: 41.2 },
+          { year: 2045, value: 382392, rate: 43.23 },
+        ],
+      },
+    ],
+  },
+  4: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1743195 },
+          { year: 1965, value: 1753126 },
+          { year: 1970, value: 1819223 },
+          { year: 1975, value: 1955267 },
+          { year: 1980, value: 2082320 },
+          { year: 1985, value: 2176295 },
+          { year: 1990, value: 2248558 },
+          { year: 1995, value: 2328739 },
+          { year: 2000, value: 2365320 },
+          { year: 2005, value: 2360218 },
+          { year: 2010, value: 2348165 },
+          { year: 2015, value: 2333899 },
+          { year: 2020, value: 2301996 },
+          { year: 2025, value: 2227471 },
+          { year: 2030, value: 2143601 },
+          { year: 2035, value: 2046219 },
+          { year: 2040, value: 1933258 },
+          { year: 2045, value: 1809021 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 584497, rate: 33.53 },
+          { year: 1965, value: 503630, rate: 28.73 },
+          { year: 1970, value: 448834, rate: 24.67 },
+          { year: 1975, value: 460953, rate: 23.57 },
+          { year: 1980, value: 480552, rate: 23.08 },
+          { year: 1985, value: 480899, rate: 22.1 },
+          { year: 1990, value: 439313, rate: 19.54 },
+          { year: 1995, value: 394331, rate: 16.93 },
+          { year: 2000, value: 353516, rate: 14.95 },
+          { year: 2005, value: 325829, rate: 13.81 },
+          { year: 2010, value: 308201, rate: 13.13 },
+          { year: 2015, value: 286003, rate: 12.25 },
+          { year: 2020, value: 268428, rate: 11.66 },
+          { year: 2025, value: 248940, rate: 11.18 },
+          { year: 2030, value: 228867, rate: 10.68 },
+          { year: 2035, value: 208564, rate: 10.19 },
+          { year: 2040, value: 191022, rate: 9.88 },
+          { year: 2045, value: 173630, rate: 9.6 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1063732, rate: 61.02 },
+          { year: 1965, value: 1141867, rate: 65.13 },
+          { year: 1970, value: 1244711, rate: 68.42 },
+          { year: 1975, value: 1343632, rate: 68.72 },
+          { year: 1980, value: 1419497, rate: 68.17 },
+          { year: 1985, value: 1478820, rate: 67.95 },
+          { year: 1990, value: 1535449, rate: 68.29 },
+          { year: 1995, value: 1595534, rate: 68.51 },
+          { year: 2000, value: 1601826, rate: 67.72 },
+          { year: 2005, value: 1558087, rate: 66.01 },
+          { year: 2010, value: 1501638, rate: 63.95 },
+          { year: 2015, value: 1410322, rate: 60.43 },
+          { year: 2020, value: 1346845, rate: 58.51 },
+          { year: 2025, value: 1282718, rate: 57.59 },
+          { year: 2030, value: 1205387, rate: 56.23 },
+          { year: 2035, value: 1121287, rate: 54.8 },
+          { year: 2040, value: 1009234, rate: 52.2 },
+          { year: 2045, value: 905996, rate: 50.08 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 94966, rate: 5.45 },
+          { year: 1965, value: 107629, rate: 6.14 },
+          { year: 1970, value: 125678, rate: 6.91 },
+          { year: 1975, value: 150010, rate: 7.67 },
+          { year: 1980, value: 180689, rate: 8.68 },
+          { year: 1985, value: 215457, rate: 9.9 },
+          { year: 1990, value: 266759, rate: 11.86 },
+          { year: 1995, value: 337520, rate: 14.49 },
+          { year: 2000, value: 409156, rate: 17.3 },
+          { year: 2005, value: 470512, rate: 19.94 },
+          { year: 2010, value: 520794, rate: 22.18 },
+          { year: 2015, value: 588240, rate: 25.2 },
+          { year: 2020, value: 638984, rate: 27.76 },
+          { year: 2025, value: 695813, rate: 31.24 },
+          { year: 2030, value: 709347, rate: 33.09 },
+          { year: 2035, value: 716368, rate: 35.01 },
+          { year: 2040, value: 733002, rate: 37.92 },
+          { year: 2045, value: 729395, rate: 40.32 },
+        ],
+      },
+    ],
+  },
+  5: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1335580 },
+          { year: 1965, value: 1279835 },
+          { year: 1970, value: 1241376 },
+          { year: 1975, value: 1232481 },
+          { year: 1980, value: 1256745 },
+          { year: 1985, value: 1254032 },
+          { year: 1990, value: 1227478 },
+          { year: 1995, value: 1213667 },
+          { year: 2000, value: 1189279 },
+          { year: 2005, value: 1145501 },
+          { year: 2010, value: 1085997 },
+          { year: 2015, value: 1023119 },
+          { year: 2020, value: 959502 },
+          { year: 2025, value: 885285 },
+          { year: 2030, value: 814295 },
+          { year: 2035, value: 744014 },
+          { year: 2040, value: 672617 },
+          { year: 2045, value: 601649 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 457929, rate: 34.29 },
+          { year: 1965, value: 373157, rate: 29.16 },
+          { year: 1970, value: 303286, rate: 24.43 },
+          { year: 1975, value: 273808, rate: 22.22 },
+          { year: 1980, value: 263277, rate: 20.95 },
+          { year: 1985, value: 250402, rate: 19.97 },
+          { year: 1990, value: 219637, rate: 17.89 },
+          { year: 1995, value: 188844, rate: 15.56 },
+          { year: 2000, value: 163095, rate: 13.71 },
+          { year: 2005, value: 142507, rate: 12.44 },
+          { year: 2010, value: 124061, rate: 11.42 },
+          { year: 2015, value: 106041, rate: 10.36 },
+          { year: 2020, value: 92673, rate: 9.66 },
+          { year: 2025, value: 79478, rate: 8.98 },
+          { year: 2030, value: 69184, rate: 8.5 },
+          { year: 2035, value: 59544, rate: 8 },
+          { year: 2040, value: 51463, rate: 7.65 },
+          { year: 2045, value: 44295, rate: 7.36 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 816077, rate: 61.1 },
+          { year: 1965, value: 832800, rate: 65.07 },
+          { year: 1970, value: 847538, rate: 68.27 },
+          { year: 1975, value: 849459, rate: 68.92 },
+          { year: 1980, value: 861312, rate: 68.54 },
+          { year: 1985, value: 845473, rate: 67.42 },
+          { year: 1990, value: 815840, rate: 66.46 },
+          { year: 1995, value: 786983, rate: 64.84 },
+          { year: 2000, value: 746252, rate: 62.75 },
+          { year: 2005, value: 694288, rate: 60.61 },
+          { year: 2010, value: 639633, rate: 58.9 },
+          { year: 2015, value: 565237, rate: 55.25 },
+          { year: 2020, value: 500687, rate: 52.18 },
+          { year: 2025, value: 444557, rate: 50.22 },
+          { year: 2030, value: 394681, rate: 48.47 },
+          { year: 2035, value: 350711, rate: 47.14 },
+          { year: 2040, value: 301533, rate: 44.83 },
+          { year: 2045, value: 255932, rate: 42.54 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 61574, rate: 4.61 },
+          { year: 1965, value: 73878, rate: 5.77 },
+          { year: 1970, value: 90552, rate: 7.29 },
+          { year: 1975, value: 109172, rate: 8.86 },
+          { year: 1980, value: 132075, rate: 10.51 },
+          { year: 1985, value: 158086, rate: 12.61 },
+          { year: 1990, value: 191573, rate: 15.61 },
+          { year: 1995, value: 237682, rate: 19.58 },
+          { year: 2000, value: 279764, rate: 23.52 },
+          { year: 2005, value: 308193, rate: 26.9 },
+          { year: 2010, value: 320450, rate: 29.51 },
+          { year: 2015, value: 343301, rate: 33.55 },
+          { year: 2020, value: 357568, rate: 37.27 },
+          { year: 2025, value: 361250, rate: 40.81 },
+          { year: 2030, value: 350430, rate: 43.03 },
+          { year: 2035, value: 333759, rate: 44.86 },
+          { year: 2040, value: 319621, rate: 47.52 },
+          { year: 2045, value: 301422, rate: 50.1 },
+        ],
+      },
+    ],
+  },
+  6: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1320664 },
+          { year: 1965, value: 1263103 },
+          { year: 1970, value: 1225618 },
+          { year: 1975, value: 1220302 },
+          { year: 1980, value: 1251917 },
+          { year: 1985, value: 1261662 },
+          { year: 1990, value: 1258390 },
+          { year: 1995, value: 1256958 },
+          { year: 2000, value: 1244147 },
+          { year: 2005, value: 1216181 },
+          { year: 2010, value: 1168924 },
+          { year: 2015, value: 1123891 },
+          { year: 2020, value: 1068027 },
+          { year: 2025, value: 1015910 },
+          { year: 2030, value: 957314 },
+          { year: 2035, value: 897075 },
+          { year: 2040, value: 833844 },
+          { year: 2045, value: 768490 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 422576, rate: 32 },
+          { year: 1965, value: 348572, rate: 27.6 },
+          { year: 1970, value: 287877, rate: 23.49 },
+          { year: 1975, value: 265935, rate: 21.79 },
+          { year: 1980, value: 262704, rate: 20.98 },
+          { year: 1985, value: 255853, rate: 20.28 },
+          { year: 1990, value: 233824, rate: 18.58 },
+          { year: 1995, value: 208596, rate: 16.6 },
+          { year: 2000, value: 186182, rate: 14.96 },
+          { year: 2005, value: 166653, rate: 13.7 },
+          { year: 2010, value: 149759, rate: 12.81 },
+          { year: 2015, value: 135760, rate: 12.08 },
+          { year: 2020, value: 120086, rate: 11.24 },
+          { year: 2025, value: 111246, rate: 10.95 },
+          { year: 2030, value: 101147, rate: 10.57 },
+          { year: 2035, value: 91199, rate: 10.17 },
+          { year: 2040, value: 82468, rate: 9.89 },
+          { year: 2045, value: 73952, rate: 9.62 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 821599, rate: 62.21 },
+          { year: 1965, value: 827075, rate: 65.48 },
+          { year: 1970, value: 833203, rate: 67.98 },
+          { year: 1975, value: 831116, rate: 68.11 },
+          { year: 1980, value: 842612, rate: 67.31 },
+          { year: 1985, value: 836219, rate: 66.28 },
+          { year: 1990, value: 819200, rate: 65.1 },
+          { year: 1995, value: 799251, rate: 63.59 },
+          { year: 2000, value: 772100, rate: 62.06 },
+          { year: 2005, value: 739030, rate: 60.77 },
+          { year: 2010, value: 694110, rate: 59.38 },
+          { year: 2015, value: 639336, rate: 56.89 },
+          { year: 2020, value: 578819, rate: 54.2 },
+          { year: 2025, value: 538804, rate: 53.04 },
+          { year: 2030, value: 496262, rate: 51.84 },
+          { year: 2035, value: 456869, rate: 50.93 },
+          { year: 2040, value: 409519, rate: 49.11 },
+          { year: 2045, value: 364262, rate: 47.4 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 76489, rate: 5.79 },
+          { year: 1965, value: 87456, rate: 6.92 },
+          { year: 1970, value: 104538, rate: 8.53 },
+          { year: 1975, value: 123137, rate: 10.09 },
+          { year: 1980, value: 146593, rate: 11.71 },
+          { year: 1985, value: 169525, rate: 13.44 },
+          { year: 1990, value: 204577, rate: 16.26 },
+          { year: 1995, value: 248817, rate: 19.8 },
+          { year: 2000, value: 285590, rate: 22.95 },
+          { year: 2005, value: 309913, rate: 25.48 },
+          { year: 2010, value: 321722, rate: 27.52 },
+          { year: 2015, value: 344353, rate: 30.64 },
+          { year: 2020, value: 359554, rate: 33.67 },
+          { year: 2025, value: 365860, rate: 36.01 },
+          { year: 2030, value: 359905, rate: 37.6 },
+          { year: 2035, value: 349007, rate: 38.9 },
+          { year: 2040, value: 341857, rate: 41 },
+          { year: 2045, value: 330276, rate: 42.98 },
+        ],
+      },
+    ],
+  },
+  7: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 2051137 },
+          { year: 1965, value: 1983754 },
+          { year: 1970, value: 1946077 },
+          { year: 1975, value: 1970616 },
+          { year: 1980, value: 2035272 },
+          { year: 1985, value: 2080304 },
+          { year: 1990, value: 2104058 },
+          { year: 1995, value: 2133592 },
+          { year: 2000, value: 2126935 },
+          { year: 2005, value: 2091319 },
+          { year: 2010, value: 2029064 },
+          { year: 2015, value: 1914039 },
+          { year: 2020, value: 1833152 },
+          { year: 2025, value: 1733103 },
+          { year: 2030, value: 1635235 },
+          { year: 2035, value: 1533521 },
+          { year: 2040, value: 1426392 },
+          { year: 2045, value: 1314903 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 720721, rate: 35.14 },
+          { year: 1965, value: 608959, rate: 30.7 },
+          { year: 1970, value: 508172, rate: 26.11 },
+          { year: 1975, value: 473903, rate: 24.05 },
+          { year: 1980, value: 466840, rate: 22.94 },
+          { year: 1985, value: 460767, rate: 22.15 },
+          { year: 1990, value: 422064, rate: 20.06 },
+          { year: 1995, value: 381511, rate: 17.88 },
+          { year: 2000, value: 341038, rate: 16.03 },
+          { year: 2005, value: 307294, rate: 14.69 },
+          { year: 2010, value: 276069, rate: 13.61 },
+          { year: 2015, value: 228887, rate: 11.96 },
+          { year: 2020, value: 206152, rate: 11.25 },
+          { year: 2025, value: 183630, rate: 10.6 },
+          { year: 2030, value: 165607, rate: 10.13 },
+          { year: 2035, value: 148595, rate: 9.69 },
+          { year: 2040, value: 134354, rate: 9.42 },
+          { year: 2045, value: 120644, rate: 9.18 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1208990, rate: 58.94 },
+          { year: 1965, value: 1239658, rate: 62.49 },
+          { year: 1970, value: 1283176, rate: 65.94 },
+          { year: 1975, value: 1316161, rate: 66.79 },
+          { year: 1980, value: 1355601, rate: 66.61 },
+          { year: 1985, value: 1371556, rate: 65.93 },
+          { year: 1990, value: 1377857, rate: 65.49 },
+          { year: 1995, value: 1380208, rate: 64.69 },
+          { year: 2000, value: 1353500, rate: 63.64 },
+          { year: 2005, value: 1307734, rate: 62.53 },
+          { year: 2010, value: 1236458, rate: 60.94 },
+          { year: 2015, value: 1120189, rate: 58.52 },
+          { year: 2020, value: 1020241, rate: 55.66 },
+          { year: 2025, value: 937556, rate: 54.1 },
+          { year: 2030, value: 856169, rate: 52.36 },
+          { year: 2035, value: 780107, rate: 50.87 },
+          { year: 2040, value: 690477, rate: 48.41 },
+          { year: 2045, value: 612508, rate: 46.58 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 121426, rate: 5.92 },
+          { year: 1965, value: 135137, rate: 6.81 },
+          { year: 1970, value: 154729, rate: 7.95 },
+          { year: 1975, value: 180356, rate: 9.15 },
+          { year: 1980, value: 212704, rate: 10.45 },
+          { year: 1985, value: 247947, rate: 11.92 },
+          { year: 1990, value: 301552, rate: 14.33 },
+          { year: 1995, value: 371572, rate: 17.42 },
+          { year: 2000, value: 431797, rate: 20.3 },
+          { year: 2005, value: 474860, rate: 22.71 },
+          { year: 2010, value: 504451, rate: 24.86 },
+          { year: 2015, value: 542384, rate: 28.34 },
+          { year: 2020, value: 572825, rate: 31.25 },
+          { year: 2025, value: 611917, rate: 35.31 },
+          { year: 2030, value: 613459, rate: 37.52 },
+          { year: 2035, value: 604819, rate: 39.44 },
+          { year: 2040, value: 601561, rate: 42.17 },
+          { year: 2045, value: 581751, rate: 44.24 },
+        ],
+      },
+    ],
+  },
+  8: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 2047024 },
+          { year: 1965, value: 2056154 },
+          { year: 1970, value: 2143551 },
+          { year: 1975, value: 2342198 },
+          { year: 1980, value: 2558007 },
+          { year: 1985, value: 2725005 },
+          { year: 1990, value: 2845382 },
+          { year: 1995, value: 2955530 },
+          { year: 2000, value: 2985676 },
+          { year: 2005, value: 2975167 },
+          { year: 2010, value: 2969770 },
+          { year: 2015, value: 2916976 },
+          { year: 2020, value: 2867009 },
+          { year: 2025, value: 2750204 },
+          { year: 2030, value: 2638185 },
+          { year: 2035, value: 2512144 },
+          { year: 2040, value: 2376146 },
+          { year: 2045, value: 2235686 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 672204, rate: 32.84 },
+          { year: 1965, value: 580171, rate: 28.22 },
+          { year: 1970, value: 534225, rate: 24.92 },
+          { year: 1975, value: 580187, rate: 24.77 },
+          { year: 1980, value: 628466, rate: 24.57 },
+          { year: 1985, value: 627512, rate: 23.03 },
+          { year: 1990, value: 559033, rate: 19.65 },
+          { year: 1995, value: 505363, rate: 17.1 },
+          { year: 2000, value: 458501, rate: 15.36 },
+          { year: 2005, value: 422913, rate: 14.21 },
+          { year: 2010, value: 399638, rate: 13.46 },
+          { year: 2015, value: 364351, rate: 12.49 },
+          { year: 2020, value: 333741, rate: 11.64 },
+          { year: 2025, value: 303436, rate: 11.03 },
+          { year: 2030, value: 279345, rate: 10.59 },
+          { year: 2035, value: 256755, rate: 10.22 },
+          { year: 2040, value: 239318, rate: 10.07 },
+          { year: 2045, value: 222074, rate: 9.93 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1240591, rate: 60.6 },
+          { year: 1965, value: 1327621, rate: 64.57 },
+          { year: 1970, value: 1440059, rate: 67.18 },
+          { year: 1975, value: 1565349, rate: 66.83 },
+          { year: 1980, value: 1692449, rate: 66.16 },
+          { year: 1985, value: 1818697, rate: 66.74 },
+          { year: 1990, value: 1943837, rate: 68.32 },
+          { year: 1995, value: 2030045, rate: 68.69 },
+          { year: 2000, value: 2030360, rate: 68 },
+          { year: 2005, value: 1974159, rate: 66.35 },
+          { year: 2010, value: 1891701, rate: 63.7 },
+          { year: 2015, value: 1747312, rate: 59.9 },
+          { year: 2020, value: 1638165, rate: 57.14 },
+          { year: 2025, value: 1567743, rate: 57 },
+          { year: 2030, value: 1476135, rate: 55.95 },
+          { year: 2035, value: 1368703, rate: 54.48 },
+          { year: 2040, value: 1229865, rate: 51.76 },
+          { year: 2045, value: 1118773, rate: 50.04 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 134229, rate: 6.56 },
+          { year: 1965, value: 148362, rate: 7.22 },
+          { year: 1970, value: 169267, rate: 7.9 },
+          { year: 1975, value: 196380, rate: 8.38 },
+          { year: 1980, value: 236485, rate: 9.24 },
+          { year: 1985, value: 278503, rate: 10.22 },
+          { year: 1990, value: 338799, rate: 11.91 },
+          { year: 1995, value: 418610, rate: 14.16 },
+          { year: 2000, value: 495693, rate: 16.6 },
+          { year: 2005, value: 576272, rate: 19.37 },
+          { year: 2010, value: 665065, rate: 22.39 },
+          { year: 2015, value: 771678, rate: 26.45 },
+          { year: 2020, value: 839907, rate: 29.3 },
+          { year: 2025, value: 879025, rate: 31.96 },
+          { year: 2030, value: 882705, rate: 33.46 },
+          { year: 2035, value: 886686, rate: 35.3 },
+          { year: 2040, value: 906963, rate: 38.17 },
+          { year: 2045, value: 894839, rate: 40.03 },
+        ],
+      },
+    ],
+  },
+  9: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1513624 },
+          { year: 1965, value: 1521656 },
+          { year: 1970, value: 1580021 },
+          { year: 1975, value: 1698003 },
+          { year: 1980, value: 1792201 },
+          { year: 1985, value: 1866066 },
+          { year: 1990, value: 1935168 },
+          { year: 1995, value: 1984390 },
+          { year: 2000, value: 2004817 },
+          { year: 2005, value: 2016631 },
+          { year: 2010, value: 2007683 },
+          { year: 2015, value: 1974255 },
+          { year: 2020, value: 1933146 },
+          { year: 2025, value: 1872842 },
+          { year: 2030, value: 1805949 },
+          { year: 2035, value: 1730414 },
+          { year: 2040, value: 1647288 },
+          { year: 2045, value: 1560619 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 510097, rate: 33.7 },
+          { year: 1965, value: 429447, rate: 28.22 },
+          { year: 1970, value: 389016, rate: 24.62 },
+          { year: 1975, value: 414123, rate: 24.39 },
+          { year: 1980, value: 433232, rate: 24.17 },
+          { year: 1985, value: 425807, rate: 22.82 },
+          { year: 1990, value: 380087, rate: 19.64 },
+          { year: 1995, value: 339253, rate: 17.1 },
+          { year: 2000, value: 306905, rate: 15.31 },
+          { year: 2005, value: 285245, rate: 14.14 },
+          { year: 2010, value: 269823, rate: 13.44 },
+          { year: 2015, value: 252836, rate: 12.81 },
+          { year: 2020, value: 227553, rate: 11.77 },
+          { year: 2025, value: 215694, rate: 11.52 },
+          { year: 2030, value: 200974, rate: 11.13 },
+          { year: 2035, value: 186959, rate: 10.8 },
+          { year: 2040, value: 175984, rate: 10.68 },
+          { year: 2045, value: 164918, rate: 10.57 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 909207, rate: 60.07 },
+          { year: 1965, value: 986774, rate: 64.85 },
+          { year: 1970, value: 1069584, rate: 67.69 },
+          { year: 1975, value: 1143083, rate: 67.32 },
+          { year: 1980, value: 1191109, rate: 66.46 },
+          { year: 1985, value: 1243861, rate: 66.66 },
+          { year: 1990, value: 1315228, rate: 67.96 },
+          { year: 1995, value: 1350635, rate: 68.06 },
+          { year: 2000, value: 1352311, rate: 67.45 },
+          { year: 2005, value: 1336513, rate: 66.27 },
+          { year: 2010, value: 1281274, rate: 63.82 },
+          { year: 2015, value: 1203616, rate: 60.97 },
+          { year: 2020, value: 1115611, rate: 57.71 },
+          { year: 2025, value: 1084830, rate: 57.92 },
+          { year: 2030, value: 1031754, rate: 57.13 },
+          { year: 2035, value: 969256, rate: 56.01 },
+          { year: 2040, value: 883575, rate: 53.64 },
+          { year: 2045, value: 812961, rate: 52.09 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 94320, rate: 6.23 },
+          { year: 1965, value: 105435, rate: 6.93 },
+          { year: 1970, value: 121421, rate: 7.68 },
+          { year: 1975, value: 140651, rate: 8.28 },
+          { year: 1980, value: 167438, rate: 9.34 },
+          { year: 1985, value: 196328, rate: 10.52 },
+          { year: 1990, value: 238505, rate: 12.32 },
+          { year: 1995, value: 292947, rate: 14.76 },
+          { year: 2000, value: 344506, rate: 17.18 },
+          { year: 2005, value: 390896, rate: 19.38 },
+          { year: 2010, value: 438196, rate: 21.83 },
+          { year: 2015, value: 508392, rate: 25.75 },
+          { year: 2020, value: 554381, rate: 28.68 },
+          { year: 2025, value: 572318, rate: 30.56 },
+          { year: 2030, value: 573221, rate: 31.74 },
+          { year: 2035, value: 574199, rate: 33.18 },
+          { year: 2040, value: 587729, rate: 35.68 },
+          { year: 2045, value: 582740, rate: 37.34 },
+        ],
+      },
+    ],
+  },
+  10: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1578476 },
+          { year: 1965, value: 1605584 },
+          { year: 1970, value: 1658909 },
+          { year: 1975, value: 1756480 },
+          { year: 1980, value: 1848562 },
+          { year: 1985, value: 1921259 },
+          { year: 1990, value: 1966265 },
+          { year: 1995, value: 2003540 },
+          { year: 2000, value: 2024852 },
+          { year: 2005, value: 2024135 },
+          { year: 2010, value: 2008068 },
+          { year: 2015, value: 1973115 },
+          { year: 2020, value: 1939110 },
+          { year: 2025, value: 1865565 },
+          { year: 2030, value: 1796233 },
+          { year: 2035, value: 1719914 },
+          { year: 2040, value: 1637642 },
+          { year: 2045, value: 1552950 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 500151, rate: 31.69 },
+          { year: 1965, value: 423874, rate: 26.4 },
+          { year: 1970, value: 397032, rate: 23.93 },
+          { year: 1975, value: 427621, rate: 24.35 },
+          { year: 1980, value: 441548, rate: 23.89 },
+          { year: 1985, value: 424829, rate: 22.11 },
+          { year: 1990, value: 368080, rate: 18.72 },
+          { year: 1995, value: 329960, rate: 16.47 },
+          { year: 2000, value: 306895, rate: 15.16 },
+          { year: 2005, value: 291995, rate: 14.43 },
+          { year: 2010, value: 275225, rate: 13.71 },
+          { year: 2015, value: 250884, rate: 12.72 },
+          { year: 2020, value: 224304, rate: 11.57 },
+          { year: 2025, value: 207172, rate: 11.11 },
+          { year: 2030, value: 191494, rate: 10.66 },
+          { year: 2035, value: 177457, rate: 10.32 },
+          { year: 2040, value: 167501, rate: 10.23 },
+          { year: 2045, value: 157536, rate: 10.14 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 981555, rate: 62.18 },
+          { year: 1965, value: 1071433, rate: 66.73 },
+          { year: 1970, value: 1131078, rate: 68.18 },
+          { year: 1975, value: 1174392, rate: 66.86 },
+          { year: 1980, value: 1222826, rate: 66.15 },
+          { year: 1985, value: 1281535, rate: 66.7 },
+          { year: 1990, value: 1340557, rate: 68.18 },
+          { year: 1995, value: 1360039, rate: 67.88 },
+          { year: 2000, value: 1346441, rate: 66.5 },
+          { year: 2005, value: 1314259, rate: 64.93 },
+          { year: 2010, value: 1251608, rate: 62.33 },
+          { year: 2015, value: 1165780, rate: 59.08 },
+          { year: 2020, value: 1096231, rate: 56.53 },
+          { year: 2025, value: 1065673, rate: 57.12 },
+          { year: 2030, value: 1009939, rate: 56.23 },
+          { year: 2035, value: 941704, rate: 54.75 },
+          { year: 2040, value: 852223, rate: 52.04 },
+          { year: 2045, value: 783531, rate: 50.45 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 96770, rate: 6.13 },
+          { year: 1965, value: 110277, rate: 6.87 },
+          { year: 1970, value: 130799, rate: 7.88 },
+          { year: 1975, value: 154371, rate: 8.79 },
+          { year: 1980, value: 184115, rate: 9.96 },
+          { year: 1985, value: 214871, rate: 11.18 },
+          { year: 1990, value: 256367, rate: 13.04 },
+          { year: 1995, value: 313425, rate: 15.64 },
+          { year: 2000, value: 367117, rate: 18.13 },
+          { year: 2005, value: 416909, rate: 20.6 },
+          { year: 2010, value: 470520, rate: 23.43 },
+          { year: 2015, value: 540026, rate: 27.37 },
+          { year: 2020, value: 576729, rate: 29.74 },
+          { year: 2025, value: 592720, rate: 31.77 },
+          { year: 2030, value: 594800, rate: 33.11 },
+          { year: 2035, value: 600753, rate: 34.93 },
+          { year: 2040, value: 617918, rate: 37.73 },
+          { year: 2045, value: 611883, rate: 39.4 },
+        ],
+      },
+    ],
+  },
+  11: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 2430871 },
+          { year: 1965, value: 3014983 },
+          { year: 1970, value: 3866472 },
+          { year: 1975, value: 4821340 },
+          { year: 1980, value: 5420480 },
+          { year: 1985, value: 5863678 },
+          { year: 1990, value: 6405319 },
+          { year: 1995, value: 6759311 },
+          { year: 2000, value: 6938006 },
+          { year: 2005, value: 7054243 },
+          { year: 2010, value: 7194556 },
+          { year: 2015, value: 7266534 },
+          { year: 2020, value: 7344765 },
+          { year: 2025, value: 7202953 },
+          { year: 2030, value: 7076167 },
+          { year: 2035, value: 6909319 },
+          { year: 2040, value: 6721414 },
+          { year: 2045, value: 6524800 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 752024, rate: 30.94 },
+          { year: 1965, value: 790961, rate: 26.23 },
+          { year: 1970, value: 1002863, rate: 25.94 },
+          { year: 1975, value: 1340136, rate: 27.8 },
+          { year: 1980, value: 1458421, rate: 26.91 },
+          { year: 1985, value: 1380180, rate: 23.54 },
+          { year: 1990, value: 1196946, rate: 18.69 },
+          { year: 1995, value: 1090395, rate: 16.13 },
+          { year: 2000, value: 1024787, rate: 14.77 },
+          { year: 2005, value: 986361, rate: 13.98 },
+          { year: 2010, value: 953668, rate: 13.26 },
+          { year: 2015, value: 910805, rate: 12.53 },
+          { year: 2020, value: 858384, rate: 11.69 },
+          { year: 2025, value: 819197, rate: 11.37 },
+          { year: 2030, value: 778699, rate: 11 },
+          { year: 2035, value: 744151, rate: 10.77 },
+          { year: 2040, value: 723391, rate: 10.76 },
+          { year: 2045, value: 698496, rate: 10.71 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1545841, rate: 63.59 },
+          { year: 1965, value: 2066337, rate: 68.54 },
+          { year: 1970, value: 2665020, rate: 68.93 },
+          { year: 1975, value: 3223138, rate: 66.85 },
+          { year: 1980, value: 3624550, rate: 66.87 },
+          { year: 1985, value: 4060573, rate: 69.25 },
+          { year: 1990, value: 4660202, rate: 72.76 },
+          { year: 1995, value: 4976945, rate: 73.63 },
+          { year: 2000, value: 5011202, rate: 72.23 },
+          { year: 2005, value: 4892253, rate: 69.35 },
+          { year: 2010, value: 4749108, rate: 66.01 },
+          { year: 2015, value: 4507174, rate: 62.03 },
+          { year: 2020, value: 4335188, rate: 59.02 },
+          { year: 2025, value: 4350217, rate: 60.39 },
+          { year: 2030, value: 4217720, rate: 59.6 },
+          { year: 2035, value: 4002341, rate: 57.93 },
+          { year: 2040, value: 3700406, rate: 55.05 },
+          { year: 2045, value: 3491511, rate: 53.51 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 133006, rate: 5.47 },
+          { year: 1965, value: 157685, rate: 5.23 },
+          { year: 1970, value: 198589, rate: 5.14 },
+          { year: 1975, value: 256014, rate: 5.31 },
+          { year: 1980, value: 333874, rate: 6.16 },
+          { year: 1985, value: 420099, rate: 7.16 },
+          { year: 1990, value: 530539, rate: 8.28 },
+          { year: 1995, value: 681172, rate: 10.08 },
+          { year: 2000, value: 889243, rate: 12.82 },
+          { year: 2005, value: 1157006, rate: 16.4 },
+          { year: 2010, value: 1464860, rate: 20.36 },
+          { year: 2015, value: 1788735, rate: 24.62 },
+          { year: 2020, value: 1934994, rate: 26.35 },
+          { year: 2025, value: 2033539, rate: 28.23 },
+          { year: 2030, value: 2079748, rate: 29.39 },
+          { year: 2035, value: 2162827, rate: 31.3 },
+          { year: 2040, value: 2297617, rate: 34.18 },
+          { year: 2045, value: 2334793, rate: 35.78 },
+        ],
+      },
+    ],
+  },
+  12: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 2306010 },
+          { year: 1965, value: 2701770 },
+          { year: 1970, value: 3366624 },
+          { year: 1975, value: 4149147 },
+          { year: 1980, value: 4735424 },
+          { year: 1985, value: 5148163 },
+          { year: 1990, value: 5555429 },
+          { year: 1995, value: 5797782 },
+          { year: 2000, value: 5926285 },
+          { year: 2005, value: 6056462 },
+          { year: 2010, value: 6216289 },
+          { year: 2015, value: 6222666 },
+          { year: 2020, value: 6284480 },
+          { year: 2025, value: 6118170 },
+          { year: 2030, value: 5985915 },
+          { year: 2035, value: 5822882 },
+          { year: 2040, value: 5645611 },
+          { year: 2045, value: 5463363 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 700195, rate: 30.36 },
+          { year: 1965, value: 701340, rate: 25.96 },
+          { year: 1970, value: 845616, rate: 25.12 },
+          { year: 1975, value: 1099590, rate: 26.5 },
+          { year: 1980, value: 1224914, rate: 25.87 },
+          { year: 1985, value: 1191814, rate: 23.15 },
+          { year: 1990, value: 1034308, rate: 18.62 },
+          { year: 1995, value: 915719, rate: 15.79 },
+          { year: 2000, value: 842534, rate: 14.22 },
+          { year: 2005, value: 819348, rate: 13.53 },
+          { year: 2010, value: 799646, rate: 12.86 },
+          { year: 2015, value: 762112, rate: 12.25 },
+          { year: 2020, value: 734496, rate: 11.69 },
+          { year: 2025, value: 674898, rate: 11.03 },
+          { year: 2030, value: 638821, rate: 10.67 },
+          { year: 2035, value: 607921, rate: 10.44 },
+          { year: 2040, value: 589262, rate: 10.44 },
+          { year: 2045, value: 568568, rate: 10.41 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1458717, rate: 63.26 },
+          { year: 1965, value: 1828708, rate: 67.69 },
+          { year: 1970, value: 2310068, rate: 68.62 },
+          { year: 1975, value: 2785401, rate: 67.13 },
+          { year: 1980, value: 3176100, rate: 67.07 },
+          { year: 1985, value: 3547459, rate: 68.91 },
+          { year: 1990, value: 3994245, rate: 71.9 },
+          { year: 1995, value: 4224738, rate: 72.87 },
+          { year: 2000, value: 4235925, rate: 71.48 },
+          { year: 2005, value: 4154600, rate: 68.6 },
+          { year: 2010, value: 4009060, rate: 64.49 },
+          { year: 2015, value: 3779812, rate: 60.74 },
+          { year: 2020, value: 3715691, rate: 59.12 },
+          { year: 2025, value: 3652524, rate: 59.7 },
+          { year: 2030, value: 3528129, rate: 58.94 },
+          { year: 2035, value: 3339827, rate: 57.36 },
+          { year: 2040, value: 3083003, rate: 54.61 },
+          { year: 2045, value: 2905996, rate: 53.19 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 147098, rate: 6.38 },
+          { year: 1965, value: 171722, rate: 6.36 },
+          { year: 1970, value: 210940, rate: 6.27 },
+          { year: 1975, value: 261205, rate: 6.3 },
+          { year: 1980, value: 330188, rate: 6.97 },
+          { year: 1985, value: 407095, rate: 7.91 },
+          { year: 1990, value: 509837, rate: 9.18 },
+          { year: 1995, value: 651789, rate: 11.24 },
+          { year: 2000, value: 837017, rate: 14.12 },
+          { year: 2005, value: 1060343, rate: 17.51 },
+          { year: 2010, value: 1320120, rate: 21.24 },
+          { year: 2015, value: 1584419, rate: 25.46 },
+          { year: 2020, value: 1699991, rate: 27.05 },
+          { year: 2025, value: 1790748, rate: 29.27 },
+          { year: 2030, value: 1818965, rate: 30.39 },
+          { year: 2035, value: 1875134, rate: 32.2 },
+          { year: 2040, value: 1973346, rate: 34.95 },
+          { year: 2045, value: 1988799, rate: 36.4 },
+        ],
+      },
+    ],
+  },
+  13: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 9683802 },
+          { year: 1965, value: 10869244 },
+          { year: 1970, value: 11408071 },
+          { year: 1975, value: 11673554 },
+          { year: 1980, value: 11618281 },
+          { year: 1985, value: 11829363 },
+          { year: 1990, value: 11855563 },
+          { year: 1995, value: 11773605 },
+          { year: 2000, value: 12064101 },
+          { year: 2005, value: 12576601 },
+          { year: 2010, value: 13159388 },
+          { year: 2015, value: 13515271 },
+          { year: 2020, value: 14047594 },
+          { year: 2025, value: 13845936 },
+          { year: 2030, value: 13882538 },
+          { year: 2035, value: 13851782 },
+          { year: 2040, value: 13758624 },
+          { year: 2045, value: 13606683 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 2249052, rate: 23.22 },
+          { year: 1965, value: 2216945, rate: 20.4 },
+          { year: 1970, value: 2400630, rate: 21.04 },
+          { year: 1975, value: 2564449, rate: 21.97 },
+          { year: 1980, value: 2393687, rate: 20.6 },
+          { year: 1985, value: 2125337, rate: 17.97 },
+          { year: 1990, value: 1727479, rate: 14.57 },
+          { year: 1995, value: 1499126, rate: 12.73 },
+          { year: 2000, value: 1420919, rate: 11.78 },
+          { year: 2005, value: 1424667, rate: 11.33 },
+          { year: 2010, value: 1477371, rate: 11.23 },
+          { year: 2015, value: 1518130, rate: 11.23 },
+          { year: 2020, value: 1566840, rate: 11.15 },
+          { year: 2025, value: 1508463, rate: 10.89 },
+          { year: 2030, value: 1471373, rate: 10.6 },
+          { year: 2035, value: 1443170, rate: 10.42 },
+          { year: 2040, value: 1432251, rate: 10.41 },
+          { year: 2045, value: 1407573, rate: 10.34 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 7067087, rate: 72.98 },
+          { year: 1965, value: 8183336, rate: 75.29 },
+          { year: 1970, value: 8416630, rate: 73.78 },
+          { year: 1975, value: 8360219, rate: 71.62 },
+          { year: 1980, value: 8308563, rate: 71.51 },
+          { year: 1985, value: 8638299, rate: 73.02 },
+          { year: 1990, value: 8790525, rate: 74.15 },
+          { year: 1995, value: 8705099, rate: 73.94 },
+          { year: 2000, value: 8685878, rate: 72 },
+          { year: 2005, value: 8695592, rate: 69.14 },
+          { year: 2010, value: 8850225, rate: 67.25 },
+          { year: 2015, value: 8734155, rate: 64.62 },
+          { year: 2020, value: 8944193, rate: 63.67 },
+          { year: 2025, value: 9065802, rate: 65.48 },
+          { year: 2030, value: 8988837, rate: 64.75 },
+          { year: 2035, value: 8734058, rate: 63.05 },
+          { year: 2040, value: 8330069, rate: 60.54 },
+          { year: 2045, value: 8023423, rate: 58.97 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 367663, rate: 3.8 },
+          { year: 1965, value: 468963, rate: 4.31 },
+          { year: 1970, value: 590811, rate: 5.18 },
+          { year: 1975, value: 731808, rate: 6.27 },
+          { year: 1980, value: 894961, rate: 7.7 },
+          { year: 1985, value: 1055850, rate: 8.93 },
+          { year: 1990, value: 1244026, rate: 10.49 },
+          { year: 1995, value: 1530695, rate: 13 },
+          { year: 2000, value: 1910456, rate: 15.84 },
+          { year: 2005, value: 2295527, rate: 18.25 },
+          { year: 2010, value: 2642231, rate: 20.08 },
+          { year: 2015, value: 3005516, rate: 22.24 },
+          { year: 2020, value: 3107822, rate: 22.12 },
+          { year: 2025, value: 3271671, rate: 23.63 },
+          { year: 2030, value: 3422328, rate: 24.65 },
+          { year: 2035, value: 3674554, rate: 26.53 },
+          { year: 2040, value: 3996304, rate: 29.05 },
+          { year: 2045, value: 4175687, rate: 30.69 },
+        ],
+      },
+    ],
+  },
+  14: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 3443176 },
+          { year: 1965, value: 4430743 },
+          { year: 1970, value: 5472247 },
+          { year: 1975, value: 6397748 },
+          { year: 1980, value: 6924348 },
+          { year: 1985, value: 7431974 },
+          { year: 1990, value: 7980391 },
+          { year: 1995, value: 8245900 },
+          { year: 2000, value: 8489974 },
+          { year: 2005, value: 8791597 },
+          { year: 2010, value: 9048331 },
+          { year: 2015, value: 9126214 },
+          { year: 2020, value: 9237337 },
+          { year: 2025, value: 9069562 },
+          { year: 2030, value: 8933474 },
+          { year: 2035, value: 8750958 },
+          { year: 2040, value: 8541016 },
+          { year: 2045, value: 8312524 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 940719, rate: 27.32 },
+          { year: 1965, value: 1035303, rate: 23.37 },
+          { year: 1970, value: 1301772, rate: 23.79 },
+          { year: 1975, value: 1632021, rate: 25.51 },
+          { year: 1980, value: 1703063, rate: 24.6 },
+          { year: 1985, value: 1595127, rate: 21.46 },
+          { year: 1990, value: 1375769, rate: 17.24 },
+          { year: 1995, value: 1231943, rate: 14.94 },
+          { year: 2000, value: 1184231, rate: 13.95 },
+          { year: 2005, value: 1184631, rate: 13.47 },
+          { year: 2010, value: 1187743, rate: 13.13 },
+          { year: 2015, value: 1140748, rate: 12.5 },
+          { year: 2020, value: 1085763, rate: 11.75 },
+          { year: 2025, value: 1028000, rate: 11.33 },
+          { year: 2030, value: 977324, rate: 10.94 },
+          { year: 2035, value: 937488, rate: 10.71 },
+          { year: 2040, value: 916593, rate: 10.73 },
+          { year: 2045, value: 891282, rate: 10.72 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 2351581, rate: 68.3 },
+          { year: 1965, value: 3201686, rate: 72.26 },
+          { year: 1970, value: 3914556, rate: 71.53 },
+          { year: 1975, value: 4424772, rate: 69.16 },
+          { year: 1980, value: 4772542, rate: 68.92 },
+          { year: 1985, value: 5277568, rate: 71.01 },
+          { year: 1990, value: 5874445, rate: 73.61 },
+          { year: 1995, value: 6098448, rate: 73.96 },
+          { year: 2000, value: 6121470, rate: 72.1 },
+          { year: 2005, value: 6088141, rate: 69.25 },
+          { year: 2010, value: 5988857, rate: 66.19 },
+          { year: 2015, value: 5744383, rate: 62.94 },
+          { year: 2020, value: 5628918, rate: 60.94 },
+          { year: 2025, value: 5617607, rate: 61.94 },
+          { year: 2030, value: 5429825, rate: 60.78 },
+          { year: 2035, value: 5129468, rate: 58.62 },
+          { year: 2040, value: 4756673, rate: 55.69 },
+          { year: 2045, value: 4498417, rate: 54.12 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 150876, rate: 4.38 },
+          { year: 1965, value: 193754, rate: 4.37 },
+          { year: 1970, value: 255919, rate: 4.68 },
+          { year: 1975, value: 337305, rate: 5.27 },
+          { year: 1980, value: 443048, rate: 6.4 },
+          { year: 1985, value: 555885, rate: 7.48 },
+          { year: 1990, value: 704596, rate: 8.83 },
+          { year: 1995, value: 908467, rate: 11.02 },
+          { year: 2000, value: 1169528, rate: 13.78 },
+          { year: 2005, value: 1480262, rate: 16.84 },
+          { year: 2010, value: 1819503, rate: 20.11 },
+          { year: 2015, value: 2158157, rate: 23.65 },
+          { year: 2020, value: 2308578, rate: 24.99 },
+          { year: 2025, value: 2423955, rate: 26.73 },
+          { year: 2030, value: 2526325, rate: 28.28 },
+          { year: 2035, value: 2684002, rate: 30.67 },
+          { year: 2040, value: 2867750, rate: 33.58 },
+          { year: 2045, value: 2922825, rate: 35.16 },
+        ],
+      },
+    ],
+  },
+  15: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 2442037 },
+          { year: 1965, value: 2398931 },
+          { year: 1970, value: 2360982 },
+          { year: 1975, value: 2391938 },
+          { year: 1980, value: 2451357 },
+          { year: 1985, value: 2478470 },
+          { year: 1990, value: 2474583 },
+          { year: 1995, value: 2488364 },
+          { year: 2000, value: 2475733 },
+          { year: 2005, value: 2431459 },
+          { year: 2010, value: 2374450 },
+          { year: 2015, value: 2304264 },
+          { year: 2020, value: 2201272 },
+          { year: 2025, value: 2131156 },
+          { year: 2030, value: 2031445 },
+          { year: 2035, value: 1926432 },
+          { year: 2040, value: 1814665 },
+          { year: 2045, value: 1698989 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 798808, rate: 32.71 },
+          { year: 1965, value: 662459, rate: 27.61 },
+          { year: 1970, value: 573414, rate: 24.29 },
+          { year: 1975, value: 557279, rate: 23.3 },
+          { year: 1980, value: 552118, rate: 22.52 },
+          { year: 1985, value: 523485, rate: 21.12 },
+          { year: 1990, value: 462559, rate: 18.69 },
+          { year: 1995, value: 406791, rate: 16.35 },
+          { year: 2000, value: 365667, rate: 14.77 },
+          { year: 2005, value: 330302, rate: 13.58 },
+          { year: 2010, value: 301708, rate: 12.71 },
+          { year: 2015, value: 275945, rate: 11.98 },
+          { year: 2020, value: 247480, rate: 11.24 },
+          { year: 2025, value: 234024, rate: 10.98 },
+          { year: 2030, value: 215212, rate: 10.59 },
+          { year: 2035, value: 197841, rate: 10.27 },
+          { year: 2040, value: 183993, rate: 10.14 },
+          { year: 2045, value: 170148, rate: 10.01 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1490414, rate: 61.03 },
+          { year: 1965, value: 1571831, rate: 65.52 },
+          { year: 1970, value: 1597283, rate: 67.65 },
+          { year: 1975, value: 1605664, rate: 67.13 },
+          { year: 1980, value: 1625404, rate: 66.31 },
+          { year: 1985, value: 1637657, rate: 66.08 },
+          { year: 1990, value: 1632329, rate: 65.96 },
+          { year: 1995, value: 1625889, rate: 65.34 },
+          { year: 2000, value: 1581186, rate: 63.87 },
+          { year: 2005, value: 1514869, rate: 62.3 },
+          { year: 2010, value: 1441262, rate: 60.7 },
+          { year: 2015, value: 1333453, rate: 57.87 },
+          { year: 2020, value: 1210917, rate: 55.01 },
+          { year: 2025, value: 1164053, rate: 54.62 },
+          { year: 2030, value: 1092613, rate: 53.79 },
+          { year: 2035, value: 1015488, rate: 52.71 },
+          { year: 2040, value: 918672, rate: 50.62 },
+          { year: 2045, value: 833902, rate: 49.08 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 152815, rate: 6.26 },
+          { year: 1965, value: 164641, rate: 6.86 },
+          { year: 1970, value: 190285, rate: 8.06 },
+          { year: 1975, value: 228759, rate: 9.56 },
+          { year: 1980, value: 273439, rate: 11.15 },
+          { year: 1985, value: 317159, rate: 12.8 },
+          { year: 1990, value: 377857, rate: 15.27 },
+          { year: 1995, value: 455064, rate: 18.29 },
+          { year: 2000, value: 526112, rate: 21.25 },
+          { year: 2005, value: 580739, rate: 23.88 },
+          { year: 2010, value: 621187, rate: 26.16 },
+          { year: 2015, value: 685085, rate: 29.73 },
+          { year: 2020, value: 715935, rate: 32.52 },
+          { year: 2025, value: 733079, rate: 34.4 },
+          { year: 2030, value: 723620, rate: 35.62 },
+          { year: 2035, value: 713103, rate: 37.02 },
+          { year: 2040, value: 712000, rate: 39.24 },
+          { year: 2045, value: 694939, rate: 40.9 },
+        ],
+      },
+    ],
+  },
+  16: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1032614 },
+          { year: 1965, value: 1025465 },
+          { year: 1970, value: 1029695 },
+          { year: 1975, value: 1070791 },
+          { year: 1980, value: 1103459 },
+          { year: 1985, value: 1118369 },
+          { year: 1990, value: 1120161 },
+          { year: 1995, value: 1123125 },
+          { year: 2000, value: 1120851 },
+          { year: 2005, value: 1111729 },
+          { year: 2010, value: 1093247 },
+          { year: 2015, value: 1066328 },
+          { year: 2020, value: 1034814 },
+          { year: 2025, value: 996442 },
+          { year: 2030, value: 954745 },
+          { year: 2035, value: 910161 },
+          { year: 2040, value: 863342 },
+          { year: 2045, value: 817398 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 307923, rate: 29.82 },
+          { year: 1965, value: 246284, rate: 24.02 },
+          { year: 1970, value: 229658, rate: 22.3 },
+          { year: 1975, value: 246399, rate: 23.01 },
+          { year: 1980, value: 249935, rate: 22.65 },
+          { year: 1985, value: 233965, rate: 20.92 },
+          { year: 1990, value: 195598, rate: 17.46 },
+          { year: 1995, value: 170045, rate: 15.14 },
+          { year: 2000, value: 157179, rate: 14.02 },
+          { year: 2005, value: 149545, rate: 13.45 },
+          { year: 2010, value: 141936, rate: 12.98 },
+          { year: 2015, value: 128848, rate: 12.08 },
+          { year: 2020, value: 115177, rate: 11.13 },
+          { year: 2025, value: 106289, rate: 10.67 },
+          { year: 2030, value: 97881, rate: 10.25 },
+          { year: 2035, value: 90394, rate: 9.93 },
+          { year: 2040, value: 85491, rate: 9.9 },
+          { year: 2045, value: 80766, rate: 9.88 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 662018, rate: 64.11 },
+          { year: 1965, value: 709713, rate: 69.21 },
+          { year: 1970, value: 716831, rate: 69.62 },
+          { year: 1975, value: 723024, rate: 67.52 },
+          { year: 1980, value: 730099, rate: 66.16 },
+          { year: 1985, value: 740728, rate: 66.23 },
+          { year: 1990, value: 754081, rate: 67.32 },
+          { year: 1995, value: 751639, rate: 66.92 },
+          { year: 2000, value: 730541, rate: 65.18 },
+          { year: 2005, value: 702924, rate: 63.23 },
+          { year: 2010, value: 662072, rate: 60.56 },
+          { year: 2015, value: 605545, rate: 56.79 },
+          { year: 2020, value: 568052, rate: 54.89 },
+          { year: 2025, value: 553324, rate: 55.53 },
+          { year: 2030, value: 525723, rate: 55.06 },
+          { year: 2035, value: 492092, rate: 54.07 },
+          { year: 2040, value: 442952, rate: 51.31 },
+          { year: 2045, value: 407488, rate: 49.85 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 62673, rate: 6.07 },
+          { year: 1965, value: 69468, rate: 6.77 },
+          { year: 1970, value: 83206, rate: 8.08 },
+          { year: 1975, value: 101265, rate: 9.46 },
+          { year: 1980, value: 123407, rate: 11.18 },
+          { year: 1985, value: 143646, rate: 12.84 },
+          { year: 1990, value: 168946, rate: 15.08 },
+          { year: 1995, value: 201320, rate: 17.92 },
+          { year: 2000, value: 232733, rate: 20.76 },
+          { year: 2005, value: 258317, rate: 23.24 },
+          { year: 2010, value: 285102, rate: 26.08 },
+          { year: 2015, value: 322899, rate: 30.28 },
+          { year: 2020, value: 333008, rate: 32.18 },
+          { year: 2025, value: 336829, rate: 33.8 },
+          { year: 2030, value: 331141, rate: 34.68 },
+          { year: 2035, value: 327675, rate: 36 },
+          { year: 2040, value: 334899, rate: 38.79 },
+          { year: 2045, value: 329144, rate: 40.27 },
+        ],
+      },
+    ],
+  },
+  17: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 973418 },
+          { year: 1965, value: 980499 },
+          { year: 1970, value: 1002420 },
+          { year: 1975, value: 1069872 },
+          { year: 1980, value: 1119304 },
+          { year: 1985, value: 1152325 },
+          { year: 1990, value: 1164628 },
+          { year: 1995, value: 1180068 },
+          { year: 2000, value: 1180977 },
+          { year: 2005, value: 1174026 },
+          { year: 2010, value: 1169788 },
+          { year: 2015, value: 1154008 },
+          { year: 2020, value: 1132526 },
+          { year: 2025, value: 1104368 },
+          { year: 2030, value: 1070727 },
+          { year: 2035, value: 1032500 },
+          { year: 2040, value: 990439 },
+          { year: 2045, value: 947918 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 294067, rate: 30.21 },
+          { year: 1965, value: 244016, rate: 24.89 },
+          { year: 1970, value: 235380, rate: 23.48 },
+          { year: 1975, value: 258131, rate: 24.13 },
+          { year: 1980, value: 265968, rate: 23.76 },
+          { year: 1985, value: 254109, rate: 22.05 },
+          { year: 1990, value: 215171, rate: 18.48 },
+          { year: 1995, value: 189741, rate: 16.08 },
+          { year: 2000, value: 175569, rate: 14.87 },
+          { year: 2005, value: 166175, rate: 14.15 },
+          { year: 2010, value: 159283, rate: 13.62 },
+          { year: 2015, value: 148372, rate: 12.86 },
+          { year: 2020, value: 137096, rate: 12.11 },
+          { year: 2025, value: 129265, rate: 11.7 },
+          { year: 2030, value: 121098, rate: 11.31 },
+          { year: 2035, value: 113993, rate: 11.04 },
+          { year: 2040, value: 109431, rate: 11.05 },
+          { year: 2045, value: 104730, rate: 11.05 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 613802, rate: 63.06 },
+          { year: 1965, value: 666139, rate: 67.94 },
+          { year: 1970, value: 685300, rate: 68.36 },
+          { year: 1975, value: 713690, rate: 66.71 },
+          { year: 1980, value: 735098, rate: 65.67 },
+          { year: 1985, value: 761172, rate: 66.06 },
+          { year: 1990, value: 786378, rate: 67.52 },
+          { year: 1995, value: 799338, rate: 67.74 },
+          { year: 2000, value: 781137, rate: 66.14 },
+          { year: 2005, value: 761257, rate: 64.84 },
+          { year: 2010, value: 725951, rate: 62.06 },
+          { year: 2015, value: 672515, rate: 58.28 },
+          { year: 2020, value: 641383, rate: 56.63 },
+          { year: 2025, value: 632363, rate: 57.26 },
+          { year: 2030, value: 606643, rate: 56.66 },
+          { year: 2035, value: 574403, rate: 55.63 },
+          { year: 2040, value: 525905, rate: 53.1 },
+          { year: 2045, value: 490392, rate: 51.73 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 65549, rate: 6.73 },
+          { year: 1965, value: 70344, rate: 7.17 },
+          { year: 1970, value: 81740, rate: 8.15 },
+          { year: 1975, value: 97825, rate: 9.14 },
+          { year: 1980, value: 117580, rate: 10.5 },
+          { year: 1985, value: 136772, rate: 11.87 },
+          { year: 1990, value: 160692, rate: 13.8 },
+          { year: 1995, value: 190905, rate: 16.18 },
+          { year: 2000, value: 219666, rate: 18.6 },
+          { year: 2005, value: 245739, rate: 20.93 },
+          { year: 2010, value: 275337, rate: 23.54 },
+          { year: 2015, value: 317151, rate: 27.48 },
+          { year: 2020, value: 333660, rate: 29.46 },
+          { year: 2025, value: 342740, rate: 31.03 },
+          { year: 2030, value: 342986, rate: 32.03 },
+          { year: 2035, value: 344104, rate: 33.33 },
+          { year: 2040, value: 355103, rate: 35.85 },
+          { year: 2045, value: 352796, rate: 37.22 },
+        ],
+      },
+    ],
+  },
+  18: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 752696 },
+          { year: 1965, value: 750557 },
+          { year: 1970, value: 744230 },
+          { year: 1975, value: 773599 },
+          { year: 1980, value: 794354 },
+          { year: 1985, value: 817633 },
+          { year: 1990, value: 823585 },
+          { year: 1995, value: 826996 },
+          { year: 2000, value: 828944 },
+          { year: 2005, value: 821592 },
+          { year: 2010, value: 806314 },
+          { year: 2015, value: 786740 },
+          { year: 2020, value: 766863 },
+          { year: 2025, value: 737898 },
+          { year: 2030, value: 709753 },
+          { year: 2035, value: 679595 },
+          { year: 2040, value: 647241 },
+          { year: 2045, value: 614144 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 230950, rate: 30.68 },
+          { year: 1965, value: 196959, rate: 26.24 },
+          { year: 1970, value: 178056, rate: 23.92 },
+          { year: 1975, value: 182670, rate: 23.61 },
+          { year: 1980, value: 181496, rate: 22.85 },
+          { year: 1985, value: 176135, rate: 21.54 },
+          { year: 1990, value: 155998, rate: 18.94 },
+          { year: 1995, value: 140593, rate: 17 },
+          { year: 2000, value: 130143, rate: 15.7 },
+          { year: 2005, value: 120745, rate: 14.7 },
+          { year: 2010, value: 112192, rate: 13.91 },
+          { year: 2015, value: 102986, rate: 13.09 },
+          { year: 2020, value: 95544, rate: 12.46 },
+          { year: 2025, value: 88480, rate: 11.99 },
+          { year: 2030, value: 82411, rate: 11.61 },
+          { year: 2035, value: 76967, rate: 11.33 },
+          { year: 2040, value: 72912, rate: 11.27 },
+          { year: 2045, value: 68833, rate: 11.21 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 467299, rate: 62.08 },
+          { year: 1965, value: 494820, rate: 65.93 },
+          { year: 1970, value: 499142, rate: 67.07 },
+          { year: 1975, value: 512478, rate: 66.25 },
+          { year: 1980, value: 521186, rate: 65.61 },
+          { year: 1985, value: 536679, rate: 65.64 },
+          { year: 1990, value: 545297, rate: 66.21 },
+          { year: 1995, value: 539592, rate: 65.25 },
+          { year: 2000, value: 529017, rate: 63.82 },
+          { year: 2005, value: 513858, rate: 62.54 },
+          { year: 2010, value: 485409, rate: 60.2 },
+          { year: 2015, value: 451409, rate: 57.38 },
+          { year: 2020, value: 427984, rate: 55.81 },
+          { year: 2025, value: 409711, rate: 55.52 },
+          { year: 2030, value: 387306, rate: 54.57 },
+          { year: 2035, value: 364792, rate: 53.68 },
+          { year: 2040, value: 333526, rate: 51.53 },
+          { year: 2045, value: 308989, rate: 50.31 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 54447, rate: 7.23 },
+          { year: 1965, value: 58778, rate: 7.83 },
+          { year: 1970, value: 67032, rate: 9.01 },
+          { year: 1975, value: 78349, rate: 10.13 },
+          { year: 1980, value: 91595, rate: 11.53 },
+          { year: 1985, value: 104794, rate: 12.82 },
+          { year: 1990, value: 121940, rate: 14.81 },
+          { year: 1995, value: 146728, rate: 17.74 },
+          { year: 2000, value: 169489, rate: 20.45 },
+          { year: 2005, value: 185501, rate: 22.58 },
+          { year: 2010, value: 200942, rate: 24.92 },
+          { year: 2015, value: 222408, rate: 28.27 },
+          { year: 2020, value: 232684, rate: 30.34 },
+          { year: 2025, value: 239707, rate: 32.49 },
+          { year: 2030, value: 240036, rate: 33.82 },
+          { year: 2035, value: 237836, rate: 35 },
+          { year: 2040, value: 240803, rate: 37.2 },
+          { year: 2045, value: 236322, rate: 38.48 },
+        ],
+      },
+    ],
+  },
+  19: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 782062 },
+          { year: 1965, value: 763194 },
+          { year: 1970, value: 762029 },
+          { year: 1975, value: 783050 },
+          { year: 1980, value: 804256 },
+          { year: 1985, value: 832832 },
+          { year: 1990, value: 852966 },
+          { year: 1995, value: 881996 },
+          { year: 2000, value: 888172 },
+          { year: 2005, value: 884515 },
+          { year: 2010, value: 863075 },
+          { year: 2015, value: 834930 },
+          { year: 2020, value: 809974 },
+          { year: 2025, value: 763274 },
+          { year: 2030, value: 724352 },
+          { year: 2035, value: 683945 },
+          { year: 2040, value: 641932 },
+          { year: 2045, value: 598935 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 246727, rate: 31.55 },
+          { year: 1965, value: 209781, rate: 27.49 },
+          { year: 1970, value: 188449, rate: 24.73 },
+          { year: 1975, value: 187047, rate: 23.89 },
+          { year: 1980, value: 182166, rate: 22.65 },
+          { year: 1985, value: 172875, rate: 20.76 },
+          { year: 1990, value: 155849, rate: 18.27 },
+          { year: 1995, value: 146048, rate: 16.56 },
+          { year: 2000, value: 137594, rate: 15.49 },
+          { year: 2005, value: 127627, rate: 14.43 },
+          { year: 2010, value: 115337, rate: 13.36 },
+          { year: 2015, value: 102270, rate: 12.25 },
+          { year: 2020, value: 91629, rate: 11.31 },
+          { year: 2025, value: 81917, rate: 10.73 },
+          { year: 2030, value: 74344, rate: 10.26 },
+          { year: 2035, value: 68082, rate: 9.95 },
+          { year: 2040, value: 63436, rate: 9.88 },
+          { year: 2045, value: 58823, rate: 9.82 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 481033, rate: 61.51 },
+          { year: 1965, value: 493846, rate: 64.71 },
+          { year: 1970, value: 505237, rate: 66.3 },
+          { year: 1975, value: 515970, rate: 65.89 },
+          { year: 1980, value: 528571, rate: 65.72 },
+          { year: 1985, value: 552259, rate: 66.31 },
+          { year: 1990, value: 569805, rate: 66.8 },
+          { year: 1995, value: 584721, rate: 66.3 },
+          { year: 2000, value: 576767, rate: 64.94 },
+          { year: 2005, value: 562495, rate: 63.59 },
+          { year: 2010, value: 531455, rate: 61.58 },
+          { year: 2015, value: 488845, rate: 58.55 },
+          { year: 2020, value: 453633, rate: 56.01 },
+          { year: 2025, value: 424027, rate: 55.55 },
+          { year: 2030, value: 389009, rate: 53.7 },
+          { year: 2035, value: 351810, rate: 51.44 },
+          { year: 2040, value: 312478, rate: 48.68 },
+          { year: 2045, value: 282742, rate: 47.21 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 54302, rate: 6.94 },
+          { year: 1965, value: 59567, rate: 7.8 },
+          { year: 1970, value: 68343, rate: 8.97 },
+          { year: 1975, value: 79870, rate: 10.2 },
+          { year: 1980, value: 93338, rate: 11.61 },
+          { year: 1985, value: 107690, rate: 12.93 },
+          { year: 1990, value: 126583, rate: 14.84 },
+          { year: 1995, value: 151148, rate: 17.14 },
+          { year: 2000, value: 173580, rate: 19.54 },
+          { year: 2005, value: 193580, rate: 21.89 },
+          { year: 2010, value: 211581, rate: 24.51 },
+          { year: 2015, value: 234544, rate: 28.09 },
+          { year: 2020, value: 245884, rate: 30.36 },
+          { year: 2025, value: 257330, rate: 33.71 },
+          { year: 2030, value: 260999, rate: 36.03 },
+          { year: 2035, value: 264053, rate: 38.61 },
+          { year: 2040, value: 266018, rate: 41.44 },
+          { year: 2045, value: 257370, rate: 42.97 },
+        ],
+      },
+    ],
+  },
+  20: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1981433 },
+          { year: 1965, value: 1958007 },
+          { year: 1970, value: 1956917 },
+          { year: 1975, value: 2017564 },
+          { year: 1980, value: 2083934 },
+          { year: 1985, value: 2136927 },
+          { year: 1990, value: 2156627 },
+          { year: 1995, value: 2193984 },
+          { year: 2000, value: 2215168 },
+          { year: 2005, value: 2196114 },
+          { year: 2010, value: 2152449 },
+          { year: 2015, value: 2098804 },
+          { year: 2020, value: 2048011 },
+          { year: 2025, value: 1957953 },
+          { year: 2030, value: 1877667 },
+          { year: 2035, value: 1793174 },
+          { year: 2040, value: 1704857 },
+          { year: 2045, value: 1614511 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 571749, rate: 28.86 },
+          { year: 1965, value: 484702, rate: 24.75 },
+          { year: 1970, value: 449533, rate: 22.97 },
+          { year: 1975, value: 464427, rate: 23.02 },
+          { year: 1980, value: 468613, rate: 22.49 },
+          { year: 1985, value: 446549, rate: 20.9 },
+          { year: 1990, value: 392889, rate: 18.22 },
+          { year: 1995, value: 355267, rate: 16.19 },
+          { year: 2000, value: 334306, rate: 15.09 },
+          { year: 2005, value: 316368, rate: 14.41 },
+          { year: 2010, value: 295742, rate: 13.74 },
+          { year: 2015, value: 269752, rate: 12.85 },
+          { year: 2020, value: 242873, rate: 11.86 },
+          { year: 2025, value: 221821, rate: 11.33 },
+          { year: 2030, value: 203671, rate: 10.85 },
+          { year: 2035, value: 188165, rate: 10.49 },
+          { year: 2040, value: 177261, rate: 10.4 },
+          { year: 2045, value: 166853, rate: 10.33 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1267103, rate: 63.95 },
+          { year: 1965, value: 1315337, rate: 67.18 },
+          { year: 1970, value: 1323665, rate: 67.64 },
+          { year: 1975, value: 1337660, rate: 66.3 },
+          { year: 1980, value: 1362192, rate: 65.37 },
+          { year: 1985, value: 1398750, rate: 65.46 },
+          { year: 1990, value: 1416125, rate: 65.66 },
+          { year: 1995, value: 1421782, rate: 64.8 },
+          { year: 2000, value: 1404575, rate: 63.41 },
+          { year: 2005, value: 1356317, rate: 61.76 },
+          { year: 2010, value: 1281683, rate: 59.55 },
+          { year: 2015, value: 1186865, rate: 56.55 },
+          { year: 2020, value: 1118429, rate: 54.61 },
+          { year: 2025, value: 1072843, rate: 54.79 },
+          { year: 2030, value: 1009579, rate: 53.77 },
+          { year: 2035, value: 936738, rate: 52.24 },
+          { year: 2040, value: 845187, rate: 49.58 },
+          { year: 2045, value: 774381, rate: 47.96 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 142581, rate: 7.2 },
+          { year: 1965, value: 157968, rate: 8.07 },
+          { year: 1970, value: 183719, rate: 9.39 },
+          { year: 1975, value: 215328, rate: 10.67 },
+          { year: 1980, value: 253120, rate: 12.15 },
+          { year: 1985, value: 291617, rate: 13.65 },
+          { year: 1990, value: 347206, rate: 16.1 },
+          { year: 1995, value: 416608, rate: 18.99 },
+          { year: 2000, value: 475127, rate: 21.45 },
+          { year: 2005, value: 521984, rate: 23.77 },
+          { year: 2010, value: 569301, rate: 26.45 },
+          { year: 2015, value: 626085, rate: 29.83 },
+          { year: 2020, value: 646942, rate: 31.59 },
+          { year: 2025, value: 663289, rate: 33.88 },
+          { year: 2030, value: 664417, rate: 35.39 },
+          { year: 2035, value: 668271, rate: 37.27 },
+          { year: 2040, value: 682409, rate: 40.03 },
+          { year: 2045, value: 673277, rate: 41.7 },
+        ],
+      },
+    ],
+  },
+  21: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1638399 },
+          { year: 1965, value: 1700365 },
+          { year: 1970, value: 1758954 },
+          { year: 1975, value: 1867978 },
+          { year: 1980, value: 1960107 },
+          { year: 1985, value: 2028536 },
+          { year: 1990, value: 2066569 },
+          { year: 1995, value: 2100315 },
+          { year: 2000, value: 2107700 },
+          { year: 2005, value: 2107226 },
+          { year: 2010, value: 2080773 },
+          { year: 2015, value: 2031903 },
+          { year: 2020, value: 1978742 },
+          { year: 2025, value: 1901131 },
+          { year: 2030, value: 1820918 },
+          { year: 2035, value: 1735195 },
+          { year: 2040, value: 1645898 },
+          { year: 2045, value: 1556529 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 488014, rate: 29.79 },
+          { year: 1965, value: 436436, rate: 25.67 },
+          { year: 1970, value: 426860, rate: 24.27 },
+          { year: 1975, value: 462981, rate: 24.79 },
+          { year: 1980, value: 470749, rate: 24.02 },
+          { year: 1985, value: 448693, rate: 22.12 },
+          { year: 1990, value: 387665, rate: 18.76 },
+          { year: 1995, value: 347733, rate: 16.56 },
+          { year: 2000, value: 322769, rate: 15.31 },
+          { year: 2005, value: 305845, rate: 14.51 },
+          { year: 2010, value: 289748, rate: 13.93 },
+          { year: 2015, value: 266998, rate: 13.14 },
+          { year: 2020, value: 240520, rate: 12.16 },
+          { year: 2025, value: 224999, rate: 11.84 },
+          { year: 2030, value: 207284, rate: 11.38 },
+          { year: 2035, value: 191755, rate: 11.05 },
+          { year: 2040, value: 180871, rate: 10.99 },
+          { year: 2045, value: 170191, rate: 10.93 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1041553, rate: 63.57 },
+          { year: 1965, value: 1144142, rate: 67.29 },
+          { year: 1970, value: 1193795, rate: 67.87 },
+          { year: 1975, value: 1244643, rate: 66.63 },
+          { year: 1980, value: 1299372, rate: 66.29 },
+          { year: 1985, value: 1359334, rate: 67.01 },
+          { year: 1990, value: 1415333, rate: 68.49 },
+          { year: 1995, value: 1430294, rate: 68.1 },
+          { year: 2000, value: 1401064, rate: 66.47 },
+          { year: 2005, value: 1357583, rate: 64.43 },
+          { year: 2010, value: 1282800, rate: 61.65 },
+          { year: 2015, value: 1185431, rate: 58.34 },
+          { year: 2020, value: 1105215, rate: 55.85 },
+          { year: 2025, value: 1072733, rate: 56.43 },
+          { year: 2030, value: 1011968, rate: 55.57 },
+          { year: 2035, value: 942992, rate: 54.35 },
+          { year: 2040, value: 851843, rate: 51.76 },
+          { year: 2045, value: 783641, rate: 50.35 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 108832, rate: 6.64 },
+          { year: 1965, value: 119787, rate: 7.04 },
+          { year: 1970, value: 138299, rate: 7.86 },
+          { year: 1975, value: 160152, rate: 8.57 },
+          { year: 1980, value: 189895, rate: 9.69 },
+          { year: 1985, value: 220397, rate: 10.86 },
+          { year: 1990, value: 262594, rate: 12.71 },
+          { year: 1995, value: 322209, rate: 15.34 },
+          { year: 2000, value: 383168, rate: 18.18 },
+          { year: 2005, value: 442124, rate: 20.98 },
+          { year: 2010, value: 499399, rate: 24 },
+          { year: 2015, value: 567571, rate: 27.93 },
+          { year: 2020, value: 593751, rate: 30.01 },
+          { year: 2025, value: 603399, rate: 31.74 },
+          { year: 2030, value: 601666, rate: 33.04 },
+          { year: 2035, value: 600448, rate: 34.6 },
+          { year: 2040, value: 613184, rate: 37.26 },
+          { year: 2045, value: 602697, rate: 38.72 },
+        ],
+      },
+    ],
+  },
+  22: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 2756271 },
+          { year: 1965, value: 2912521 },
+          { year: 1970, value: 3089895 },
+          { year: 1975, value: 3308799 },
+          { year: 1980, value: 3446804 },
+          { year: 1985, value: 3574692 },
+          { year: 1990, value: 3670840 },
+          { year: 1995, value: 3737689 },
+          { year: 2000, value: 3767393 },
+          { year: 2005, value: 3792377 },
+          { year: 2010, value: 3765007 },
+          { year: 2015, value: 3700305 },
+          { year: 2020, value: 3633202 },
+          { year: 2025, value: 3506064 },
+          { year: 2030, value: 3380104 },
+          { year: 2035, value: 3241803 },
+          { year: 2040, value: 3094264 },
+          { year: 2045, value: 2942865 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 851953, rate: 30.91 },
+          { year: 1965, value: 776319, rate: 26.65 },
+          { year: 1970, value: 765233, rate: 24.77 },
+          { year: 1975, value: 825424, rate: 24.95 },
+          { year: 1980, value: 832816, rate: 24.16 },
+          { year: 1985, value: 795418, rate: 22.25 },
+          { year: 1990, value: 694558, rate: 18.92 },
+          { year: 1995, value: 619445, rate: 16.57 },
+          { year: 2000, value: 568986, rate: 15.1 },
+          { year: 2005, value: 536799, rate: 14.15 },
+          { year: 2010, value: 511575, rate: 13.59 },
+          { year: 2015, value: 478084, rate: 12.92 },
+          { year: 2020, value: 438723, rate: 12.08 },
+          { year: 2025, value: 407749, rate: 11.63 },
+          { year: 2030, value: 378716, rate: 11.2 },
+          { year: 2035, value: 353742, rate: 10.91 },
+          { year: 2040, value: 336049, rate: 10.86 },
+          { year: 2045, value: 318376, rate: 10.82 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1741510, rate: 63.18 },
+          { year: 1965, value: 1950298, rate: 66.96 },
+          { year: 1970, value: 2106471, rate: 68.17 },
+          { year: 1975, value: 2222518, rate: 67.17 },
+          { year: 1980, value: 2301206, rate: 66.76 },
+          { year: 1985, value: 2411753, rate: 67.47 },
+          { year: 1990, value: 2530257, rate: 68.93 },
+          { year: 1995, value: 2564385, rate: 68.61 },
+          { year: 2000, value: 2532256, rate: 67.22 },
+          { year: 2005, value: 2471335, rate: 65.17 },
+          { year: 2010, value: 2339915, rate: 62.15 },
+          { year: 2015, value: 2175004, rate: 58.78 },
+          { year: 2020, value: 2071258, rate: 57.01 },
+          { year: 2025, value: 1979377, rate: 56.46 },
+          { year: 2030, value: 1876560, rate: 55.52 },
+          { year: 2035, value: 1754449, rate: 54.12 },
+          { year: 2040, value: 1597414, rate: 51.63 },
+          { year: 2045, value: 1481066, rate: 50.33 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 162808, rate: 5.91 },
+          { year: 1965, value: 185904, rate: 6.38 },
+          { year: 1970, value: 218191, rate: 7.06 },
+          { year: 1975, value: 260626, rate: 7.88 },
+          { year: 1980, value: 312527, rate: 9.07 },
+          { year: 1985, value: 367136, rate: 10.27 },
+          { year: 1990, value: 444899, rate: 12.12 },
+          { year: 1995, value: 553530, rate: 14.81 },
+          { year: 2000, value: 665574, rate: 17.67 },
+          { year: 2005, value: 779193, rate: 20.55 },
+          { year: 2010, value: 891807, rate: 23.69 },
+          { year: 2015, value: 1021283, rate: 27.6 },
+          { year: 2020, value: 1084282, rate: 29.84 },
+          { year: 2025, value: 1118938, rate: 31.91 },
+          { year: 2030, value: 1124828, rate: 33.28 },
+          { year: 2035, value: 1133612, rate: 34.97 },
+          { year: 2040, value: 1160801, rate: 37.51 },
+          { year: 2045, value: 1143423, rate: 38.85 },
+        ],
+      },
+    ],
+  },
+  23: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 4206313 },
+          { year: 1965, value: 4798653 },
+          { year: 1970, value: 5386163 },
+          { year: 1975, value: 5923569 },
+          { year: 1980, value: 6221638 },
+          { year: 1985, value: 6455172 },
+          { year: 1990, value: 6690603 },
+          { year: 1995, value: 6868336 },
+          { year: 2000, value: 7043300 },
+          { year: 2005, value: 7254704 },
+          { year: 2010, value: 7410719 },
+          { year: 2015, value: 7483128 },
+          { year: 2020, value: 7542415 },
+          { year: 2025, value: 7455615 },
+          { year: 2030, value: 7359302 },
+          { year: 2035, value: 7227958 },
+          { year: 2040, value: 7070766 },
+          { year: 2045, value: 6899465 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 1146564, rate: 27.26 },
+          { year: 1965, value: 1149974, rate: 23.96 },
+          { year: 1970, value: 1309642, rate: 24.31 },
+          { year: 1975, value: 1536956, rate: 25.95 },
+          { year: 1980, value: 1555359, rate: 25 },
+          { year: 1985, value: 1446729, rate: 22.41 },
+          { year: 1990, value: 1236783, rate: 18.49 },
+          { year: 1995, value: 1120992, rate: 16.32 },
+          { year: 2000, value: 1081280, rate: 15.35 },
+          { year: 2005, value: 1069498, rate: 14.74 },
+          { year: 2010, value: 1065254, rate: 14.37 },
+          { year: 2015, value: 1022532, rate: 13.66 },
+          { year: 2020, value: 973642, rate: 12.91 },
+          { year: 2025, value: 931756, rate: 12.5 },
+          { year: 2030, value: 890816, rate: 12.1 },
+          { year: 2035, value: 856395, rate: 11.85 },
+          { year: 2040, value: 836416, rate: 11.83 },
+          { year: 2045, value: 811506, rate: 11.76 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 2840677, rate: 67.53 },
+          { year: 1965, value: 3394933, rate: 70.75 },
+          { year: 1970, value: 3768879, rate: 69.97 },
+          { year: 1975, value: 4010257, rate: 67.7 },
+          { year: 1980, value: 4202305, rate: 67.54 },
+          { year: 1985, value: 4458167, rate: 69.06 },
+          { year: 1990, value: 4784821, rate: 71.52 },
+          { year: 1995, value: 4919095, rate: 71.62 },
+          { year: 2000, value: 4914857, rate: 69.78 },
+          { year: 2005, value: 4901072, rate: 67.56 },
+          { year: 2010, value: 4791445, rate: 64.66 },
+          { year: 2015, value: 4618657, rate: 61.72 },
+          { year: 2020, value: 4502713, rate: 59.7 },
+          { year: 2025, value: 4574101, rate: 61.35 },
+          { year: 2030, value: 4462897, rate: 60.64 },
+          { year: 2035, value: 4278224, rate: 59.19 },
+          { year: 2040, value: 3996599, rate: 56.52 },
+          { year: 2045, value: 3803026, rate: 55.12 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 219072, rate: 5.21 },
+          { year: 1965, value: 253746, rate: 5.29 },
+          { year: 1970, value: 307642, rate: 5.71 },
+          { year: 1975, value: 375639, rate: 6.34 },
+          { year: 1980, value: 462213, rate: 7.43 },
+          { year: 1985, value: 547223, rate: 8.48 },
+          { year: 1990, value: 656283, rate: 9.81 },
+          { year: 1995, value: 819026, rate: 11.92 },
+          { year: 2000, value: 1019999, rate: 14.48 },
+          { year: 2005, value: 1248562, rate: 17.21 },
+          { year: 2010, value: 1492085, rate: 20.13 },
+          { year: 2015, value: 1760763, rate: 23.53 },
+          { year: 2020, value: 1864603, rate: 24.72 },
+          { year: 2025, value: 1949758, rate: 26.15 },
+          { year: 2030, value: 2005589, rate: 27.25 },
+          { year: 2035, value: 2093339, rate: 28.96 },
+          { year: 2040, value: 2237751, rate: 31.65 },
+          { year: 2045, value: 2284933, rate: 33.12 },
+        ],
+      },
+    ],
+  },
+  24: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1485054 },
+          { year: 1965, value: 1514467 },
+          { year: 1970, value: 1543083 },
+          { year: 1975, value: 1626002 },
+          { year: 1980, value: 1686936 },
+          { year: 1985, value: 1747311 },
+          { year: 1990, value: 1792514 },
+          { year: 1995, value: 1841358 },
+          { year: 2000, value: 1857339 },
+          { year: 2005, value: 1866963 },
+          { year: 2010, value: 1854724 },
+          { year: 2015, value: 1815865 },
+          { year: 2020, value: 1770254 },
+          { year: 2025, value: 1709820 },
+          { year: 2030, value: 1645050 },
+          { year: 2035, value: 1575867 },
+          { year: 2040, value: 1503635 },
+          { year: 2045, value: 1430804 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 427532, rate: 28.79 },
+          { year: 1965, value: 375217, rate: 24.78 },
+          { year: 1970, value: 360446, rate: 23.36 },
+          { year: 1975, value: 385615, rate: 23.72 },
+          { year: 1980, value: 385969, rate: 22.88 },
+          { year: 1985, value: 371893, rate: 21.28 },
+          { year: 1990, value: 330251, rate: 18.42 },
+          { year: 1995, value: 303645, rate: 16.49 },
+          { year: 2000, value: 283081, rate: 15.24 },
+          { year: 2005, value: 266741, rate: 14.29 },
+          { year: 2010, value: 253174, rate: 13.65 },
+          { year: 2015, value: 233525, rate: 12.86 },
+          { year: 2020, value: 211090, rate: 11.92 },
+          { year: 2025, value: 199047, rate: 11.64 },
+          { year: 2030, value: 185178, rate: 11.26 },
+          { year: 2035, value: 172945, rate: 10.97 },
+          { year: 2040, value: 164327, rate: 10.93 },
+          { year: 2045, value: 155645, rate: 10.88 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 950531, rate: 64.01 },
+          { year: 1965, value: 1019994, rate: 67.35 },
+          { year: 1970, value: 1044451, rate: 67.69 },
+          { year: 1975, value: 1080115, rate: 66.43 },
+          { year: 1980, value: 1113812, rate: 66.03 },
+          { year: 1985, value: 1164508, rate: 66.65 },
+          { year: 1990, value: 1218368, rate: 67.97 },
+          { year: 1995, value: 1240428, rate: 67.36 },
+          { year: 2000, value: 1222594, rate: 65.83 },
+          { year: 2005, value: 1197255, rate: 64.13 },
+          { year: 2010, value: 1142275, rate: 61.59 },
+          { year: 2015, value: 1061577, rate: 58.46 },
+          { year: 2020, value: 997052, rate: 56.32 },
+          { year: 2025, value: 976566, rate: 57.12 },
+          { year: 2030, value: 923306, rate: 56.13 },
+          { year: 2035, value: 863977, rate: 54.83 },
+          { year: 2040, value: 785011, rate: 52.21 },
+          { year: 2045, value: 727862, rate: 50.87 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 106991, rate: 7.2 },
+          { year: 1965, value: 119256, rate: 7.87 },
+          { year: 1970, value: 138186, rate: 8.96 },
+          { year: 1975, value: 160166, rate: 9.85 },
+          { year: 1980, value: 187019, rate: 11.09 },
+          { year: 1985, value: 210815, rate: 12.07 },
+          { year: 1990, value: 243358, rate: 13.58 },
+          { year: 1995, value: 297129, rate: 16.14 },
+          { year: 2000, value: 350959, rate: 18.9 },
+          { year: 2005, value: 400647, rate: 21.46 },
+          { year: 2010, value: 447103, rate: 24.11 },
+          { year: 2015, value: 501046, rate: 27.59 },
+          { year: 2020, value: 522073, rate: 29.49 },
+          { year: 2025, value: 534207, rate: 31.24 },
+          { year: 2030, value: 536566, rate: 32.62 },
+          { year: 2035, value: 538945, rate: 34.2 },
+          { year: 2040, value: 554297, rate: 36.86 },
+          { year: 2045, value: 547297, rate: 38.25 },
+        ],
+      },
+    ],
+  },
+  25: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 842695 },
+          { year: 1965, value: 853385 },
+          { year: 1970, value: 889768 },
+          { year: 1975, value: 985621 },
+          { year: 1980, value: 1079898 },
+          { year: 1985, value: 1155844 },
+          { year: 1990, value: 1222411 },
+          { year: 1995, value: 1287005 },
+          { year: 2000, value: 1342832 },
+          { year: 2005, value: 1380361 },
+          { year: 2010, value: 1410777 },
+          { year: 2015, value: 1412916 },
+          { year: 2020, value: 1413610 },
+          { year: 2025, value: 1394593 },
+          { year: 2030, value: 1371841 },
+          { year: 2035, value: 1341440 },
+          { year: 2040, value: 1304201 },
+          { year: 2045, value: 1262924 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 242327, rate: 28.76 },
+          { year: 1965, value: 211318, rate: 24.76 },
+          { year: 1970, value: 205440, rate: 23.09 },
+          { year: 1975, value: 238315, rate: 24.18 },
+          { year: 1980, value: 265737, rate: 24.61 },
+          { year: 1985, value: 270330, rate: 23.39 },
+          { year: 1990, value: 249258, rate: 20.39 },
+          { year: 1995, value: 231022, rate: 17.95 },
+          { year: 2000, value: 220072, rate: 16.39 },
+          { year: 2005, value: 213147, rate: 15.44 },
+          { year: 2010, value: 210753, rate: 14.94 },
+          { year: 2015, value: 203450, rate: 14.4 },
+          { year: 2020, value: 191369, rate: 13.54 },
+          { year: 2025, value: 186138, rate: 13.35 },
+          { year: 2030, value: 177370, rate: 12.93 },
+          { year: 2035, value: 169337, rate: 12.62 },
+          { year: 2040, value: 164002, rate: 12.57 },
+          { year: 2045, value: 158223, rate: 12.53 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 536713, rate: 63.69 },
+          { year: 1965, value: 572816, rate: 67.12 },
+          { year: 1970, value: 605063, rate: 68 },
+          { year: 1975, value: 655159, rate: 66.47 },
+          { year: 1980, value: 705815, rate: 65.36 },
+          { year: 1985, value: 760695, rate: 65.81 },
+          { year: 1990, value: 824232, rate: 67.43 },
+          { year: 1995, value: 874455, rate: 67.94 },
+          { year: 2000, value: 906629, rate: 67.52 },
+          { year: 2005, value: 916572, rate: 66.4 },
+          { year: 2010, value: 897583, rate: 63.62 },
+          { year: 2015, value: 857720, rate: 60.71 },
+          { year: 2020, value: 824781, rate: 58.35 },
+          { year: 2025, value: 825325, rate: 59.18 },
+          { year: 2030, value: 800961, rate: 58.39 },
+          { year: 2035, value: 767417, rate: 57.21 },
+          { year: 2040, value: 713324, rate: 54.69 },
+          { year: 2045, value: 671449, rate: 53.17 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 63655, rate: 7.55 },
+          { year: 1965, value: 69251, rate: 8.11 },
+          { year: 1970, value: 79265, rate: 8.91 },
+          { year: 1975, value: 91937, rate: 9.33 },
+          { year: 1980, value: 108245, rate: 10.02 },
+          { year: 1985, value: 124657, rate: 10.78 },
+          { year: 1990, value: 147144, rate: 12.04 },
+          { year: 1995, value: 181376, rate: 14.09 },
+          { year: 2000, value: 215552, rate: 16.05 },
+          { year: 2005, value: 249418, rate: 18.07 },
+          { year: 2010, value: 288788, rate: 20.47 },
+          { year: 2015, value: 337877, rate: 23.91 },
+          { year: 2020, value: 365311, rate: 25.84 },
+          { year: 2025, value: 383130, rate: 27.47 },
+          { year: 2030, value: 393510, rate: 28.68 },
+          { year: 2035, value: 404686, rate: 30.17 },
+          { year: 2040, value: 426875, rate: 32.73 },
+          { year: 2045, value: 433252, rate: 34.31 },
+        ],
+      },
+    ],
+  },
+  26: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1993403 },
+          { year: 1965, value: 2102808 },
+          { year: 1970, value: 2250087 },
+          { year: 1975, value: 2424856 },
+          { year: 1980, value: 2527330 },
+          { year: 1985, value: 2586574 },
+          { year: 1990, value: 2602460 },
+          { year: 1995, value: 2629592 },
+          { year: 2000, value: 2644391 },
+          { year: 2005, value: 2647660 },
+          { year: 2010, value: 2636092 },
+          { year: 2015, value: 2610353 },
+          { year: 2020, value: 2578087 },
+          { year: 2025, value: 2509875 },
+          { year: 2030, value: 2430849 },
+          { year: 2035, value: 2338843 },
+          { year: 2040, value: 2238226 },
+          { year: 2045, value: 2136807 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 510521, rate: 25.61 },
+          { year: 1965, value: 451185, rate: 21.46 },
+          { year: 1970, value: 484024, rate: 21.51 },
+          { year: 1975, value: 559934, rate: 23.09 },
+          { year: 1980, value: 575948, rate: 22.79 },
+          { year: 1985, value: 538628, rate: 20.82 },
+          { year: 1990, value: 448900, rate: 17.25 },
+          { year: 1995, value: 390138, rate: 14.84 },
+          { year: 2000, value: 360531, rate: 13.63 },
+          { year: 2005, value: 345071, rate: 13.03 },
+          { year: 2010, value: 334444, rate: 12.69 },
+          { year: 2015, value: 313866, rate: 12.02 },
+          { year: 2020, value: 293465, rate: 11.38 },
+          { year: 2025, value: 273718, rate: 10.91 },
+          { year: 2030, value: 255327, rate: 10.5 },
+          { year: 2035, value: 238908, rate: 10.21 },
+          { year: 2040, value: 227606, rate: 10.17 },
+          { year: 2045, value: 215872, rate: 10.1 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1356441, rate: 68.05 },
+          { year: 1965, value: 1503715, rate: 71.51 },
+          { year: 1970, value: 1588076, rate: 70.58 },
+          { year: 1975, value: 1646671, rate: 67.91 },
+          { year: 1980, value: 1693183, rate: 66.99 },
+          { year: 1985, value: 1757517, rate: 67.95 },
+          { year: 1990, value: 1816015, rate: 69.78 },
+          { year: 1995, value: 1842467, rate: 70.07 },
+          { year: 2000, value: 1810233, rate: 68.46 },
+          { year: 2005, value: 1755447, rate: 66.3 },
+          { year: 2010, value: 1653812, rate: 62.74 },
+          { year: 2015, value: 1539540, rate: 58.98 },
+          { year: 2020, value: 1467216, rate: 56.91 },
+          { year: 2025, value: 1474453, rate: 58.75 },
+          { year: 2030, value: 1409564, rate: 57.99 },
+          { year: 2035, value: 1322507, rate: 56.55 },
+          { year: 2040, value: 1203061, rate: 53.75 },
+          { year: 2045, value: 1113454, rate: 52.11 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 126441, rate: 6.34 },
+          { year: 1965, value: 147908, rate: 7.03 },
+          { year: 1970, value: 177987, rate: 7.91 },
+          { year: 1975, value: 217137, rate: 8.95 },
+          { year: 1980, value: 257836, rate: 10.2 },
+          { year: 1985, value: 289629, rate: 11.2 },
+          { year: 1990, value: 327429, rate: 12.58 },
+          { year: 1995, value: 386976, rate: 14.72 },
+          { year: 2000, value: 459273, rate: 17.37 },
+          { year: 2005, value: 530350, rate: 20.03 },
+          { year: 2010, value: 605709, rate: 22.98 },
+          { year: 2015, value: 703419, rate: 26.95 },
+          { year: 2020, value: 734493, rate: 28.49 },
+          { year: 2025, value: 761704, rate: 30.35 },
+          { year: 2030, value: 765958, rate: 31.51 },
+          { year: 2035, value: 777428, rate: 33.24 },
+          { year: 2040, value: 807559, rate: 36.08 },
+          { year: 2045, value: 807481, rate: 37.79 },
+        ],
+      },
+    ],
+  },
+  27: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 5504746 },
+          { year: 1965, value: 6657189 },
+          { year: 1970, value: 7620480 },
+          { year: 1975, value: 8278925 },
+          { year: 1980, value: 8473446 },
+          { year: 1985, value: 8668095 },
+          { year: 1990, value: 8734516 },
+          { year: 1995, value: 8797268 },
+          { year: 2000, value: 8805081 },
+          { year: 2005, value: 8817166 },
+          { year: 2010, value: 8865245 },
+          { year: 2015, value: 8839469 },
+          { year: 2020, value: 8837685 },
+          { year: 2025, value: 8526202 },
+          { year: 2030, value: 8262029 },
+          { year: 2035, value: 7962983 },
+          { year: 2040, value: 7649229 },
+          { year: 2045, value: 7335352 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 1408751, rate: 25.59 },
+          { year: 1965, value: 1526823, rate: 22.93 },
+          { year: 1970, value: 1820583, rate: 23.89 },
+          { year: 1975, value: 2120992, rate: 25.62 },
+          { year: 1980, value: 2067583, rate: 24.4 },
+          { year: 1985, value: 1850179, rate: 21.34 },
+          { year: 1990, value: 1503885, rate: 17.22 },
+          { year: 1995, value: 1321475, rate: 15.02 },
+          { year: 2000, value: 1249955, rate: 14.2 },
+          { year: 2005, value: 1211257, rate: 13.74 },
+          { year: 2010, value: 1165200, rate: 13.14 },
+          { year: 2015, value: 1093111, rate: 12.37 },
+          { year: 2020, value: 1029499, rate: 11.65 },
+          { year: 2025, value: 950018, rate: 11.14 },
+          { year: 2030, value: 887882, rate: 10.75 },
+          { year: 2035, value: 836296, rate: 10.5 },
+          { year: 2040, value: 803747, rate: 10.51 },
+          { year: 2045, value: 767595, rate: 10.46 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 3861837, rate: 70.15 },
+          { year: 1965, value: 4826846, rate: 72.51 },
+          { year: 1970, value: 5406092, rate: 70.94 },
+          { year: 1975, value: 5652511, rate: 68.28 },
+          { year: 1980, value: 5783300, rate: 68.25 },
+          { year: 1985, value: 6093737, rate: 70.3 },
+          { year: 1990, value: 6347525, rate: 72.67 },
+          { year: 1995, value: 6411945, rate: 72.89 },
+          { year: 2000, value: 6224186, rate: 70.69 },
+          { year: 2005, value: 5913558, rate: 67.07 },
+          { year: 2010, value: 5648070, rate: 63.71 },
+          { year: 2015, value: 5341654, rate: 60.43 },
+          { year: 2020, value: 5199504, rate: 58.83 },
+          { year: 2025, value: 5148449, rate: 60.38 },
+          { year: 2030, value: 4929450, rate: 59.66 },
+          { year: 2035, value: 4607787, rate: 57.87 },
+          { year: 2040, value: 4192275, rate: 54.81 },
+          { year: 2045, value: 3910455, rate: 53.31 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 234158, rate: 4.25 },
+          { year: 1965, value: 303520, rate: 4.56 },
+          { year: 1970, value: 393805, rate: 5.17 },
+          { year: 1975, value: 500674, rate: 6.05 },
+          { year: 1980, value: 613361, rate: 7.24 },
+          { year: 1985, value: 716579, rate: 8.27 },
+          { year: 1990, value: 843024, rate: 9.65 },
+          { year: 1995, value: 1047875, rate: 11.91 },
+          { year: 2000, value: 1315213, rate: 14.94 },
+          { year: 2005, value: 1634218, rate: 18.53 },
+          { year: 2010, value: 1962748, rate: 22.14 },
+          { year: 2015, value: 2278324, rate: 25.77 },
+          { year: 2020, value: 2361723, rate: 26.72 },
+          { year: 2025, value: 2427735, rate: 28.47 },
+          { year: 2030, value: 2444697, rate: 29.59 },
+          { year: 2035, value: 2518900, rate: 31.63 },
+          { year: 2040, value: 2653207, rate: 34.69 },
+          { year: 2045, value: 2657302, rate: 36.23 },
+        ],
+      },
+    ],
+  },
+  28: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 3906487 },
+          { year: 1965, value: 4309944 },
+          { year: 1970, value: 4667928 },
+          { year: 1975, value: 4992140 },
+          { year: 1980, value: 5144892 },
+          { year: 1985, value: 5278050 },
+          { year: 1990, value: 5405040 },
+          { year: 1995, value: 5401877 },
+          { year: 2000, value: 5550574 },
+          { year: 2005, value: 5590601 },
+          { year: 2010, value: 5588133 },
+          { year: 2015, value: 5534800 },
+          { year: 2020, value: 5465002 },
+          { year: 2025, value: 5306083 },
+          { year: 2030, value: 5139095 },
+          { year: 2035, value: 4948778 },
+          { year: 2040, value: 4742647 },
+          { year: 2045, value: 4532499 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 1089072, rate: 27.88 },
+          { year: 1965, value: 1037393, rate: 24.07 },
+          { year: 1970, value: 1096958, rate: 23.5 },
+          { year: 1975, value: 1224538, rate: 24.53 },
+          { year: 1980, value: 1227770, rate: 23.86 },
+          { year: 1985, value: 1149105, rate: 21.77 },
+          { year: 1990, value: 991045, rate: 18.34 },
+          { year: 1995, value: 880094, rate: 16.29 },
+          { year: 2000, value: 830112, rate: 14.96 },
+          { year: 2005, value: 793885, rate: 14.2 },
+          { year: 2010, value: 759277, rate: 13.59 },
+          { year: 2015, value: 706871, rate: 12.77 },
+          { year: 2020, value: 660205, rate: 12.08 },
+          { year: 2025, value: 608288, rate: 11.46 },
+          { year: 2030, value: 563409, rate: 10.96 },
+          { year: 2035, value: 525144, rate: 10.61 },
+          { year: 2040, value: 499853, rate: 10.54 },
+          { year: 2045, value: 474469, rate: 10.47 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 2594822, rate: 66.42 },
+          { year: 1965, value: 3006974, rate: 69.77 },
+          { year: 1970, value: 3246965, rate: 69.56 },
+          { year: 1975, value: 3369577, rate: 67.5 },
+          { year: 1980, value: 3435027, rate: 66.77 },
+          { year: 1985, value: 3581543, rate: 67.86 },
+          { year: 1990, value: 3752880, rate: 69.43 },
+          { year: 1995, value: 3755500, rate: 69.52 },
+          { year: 2000, value: 3776483, rate: 68.04 },
+          { year: 2005, value: 3667475, rate: 65.6 },
+          { year: 2010, value: 3515442, rate: 62.91 },
+          { year: 2015, value: 3280212, rate: 59.27 },
+          { year: 2020, value: 3075863, rate: 56.28 },
+          { year: 2025, value: 3064176, rate: 57.75 },
+          { year: 2030, value: 2916530, rate: 56.75 },
+          { year: 2035, value: 2725857, rate: 55.08 },
+          { year: 2040, value: 2472326, rate: 52.13 },
+          { year: 2045, value: 2293740, rate: 50.61 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 222593, rate: 5.7 },
+          { year: 1965, value: 265577, rate: 6.16 },
+          { year: 1970, value: 324005, rate: 6.94 },
+          { year: 1975, value: 395727, rate: 7.93 },
+          { year: 1980, value: 474708, rate: 9.23 },
+          { year: 1985, value: 545382, rate: 10.33 },
+          { year: 1990, value: 642401, rate: 11.89 },
+          { year: 1995, value: 763752, rate: 14.14 },
+          { year: 2000, value: 939950, rate: 16.93 },
+          { year: 2005, value: 1108564, rate: 19.83 },
+          { year: 2010, value: 1281486, rate: 22.93 },
+          { year: 2015, value: 1481646, rate: 26.77 },
+          { year: 2020, value: 1546543, rate: 28.3 },
+          { year: 2025, value: 1633619, rate: 30.79 },
+          { year: 2030, value: 1659156, rate: 32.28 },
+          { year: 2035, value: 1697777, rate: 34.31 },
+          { year: 2040, value: 1770468, rate: 37.33 },
+          { year: 2045, value: 1764290, rate: 38.93 },
+        ],
+      },
+    ],
+  },
+  29: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 781058 },
+          { year: 1965, value: 825965 },
+          { year: 1970, value: 930160 },
+          { year: 1975, value: 1077491 },
+          { year: 1980, value: 1209365 },
+          { year: 1985, value: 1304866 },
+          { year: 1990, value: 1375481 },
+          { year: 1995, value: 1430862 },
+          { year: 2000, value: 1442795 },
+          { year: 2005, value: 1421310 },
+          { year: 2010, value: 1400728 },
+          { year: 2015, value: 1364316 },
+          { year: 2020, value: 1324473 },
+          { year: 2025, value: 1264574 },
+          { year: 2030, value: 1202479 },
+          { year: 2035, value: 1135578 },
+          { year: 2040, value: 1066267 },
+          { year: 2045, value: 998076 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 213183, rate: 27.29 },
+          { year: 1965, value: 196950, rate: 23.84 },
+          { year: 1970, value: 212847, rate: 22.88 },
+          { year: 1975, value: 261132, rate: 24.24 },
+          { year: 1980, value: 291486, rate: 24.1 },
+          { year: 1985, value: 288515, rate: 22.11 },
+          { year: 1990, value: 255863, rate: 18.6 },
+          { year: 1995, value: 232418, rate: 16.24 },
+          { year: 2000, value: 213822, rate: 14.82 },
+          { year: 2005, value: 197136, rate: 13.87 },
+          { year: 2010, value: 184011, rate: 13.14 },
+          { year: 2015, value: 168970, rate: 12.38 },
+          { year: 2020, value: 154271, rate: 11.65 },
+          { year: 2025, value: 141163, rate: 11.16 },
+          { year: 2030, value: 128830, rate: 10.71 },
+          { year: 2035, value: 118608, rate: 10.44 },
+          { year: 2040, value: 111441, rate: 10.45 },
+          { year: 2045, value: 104315, rate: 10.45 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 515426, rate: 65.99 },
+          { year: 1965, value: 568411, rate: 68.82 },
+          { year: 1970, value: 643163, rate: 69.15 },
+          { year: 1975, value: 723989, rate: 67.19 },
+          { year: 1980, value: 804308, rate: 66.51 },
+          { year: 1985, value: 883787, rate: 67.73 },
+          { year: 1990, value: 958469, rate: 69.68 },
+          { year: 1995, value: 999261, rate: 69.84 },
+          { year: 2000, value: 987435, rate: 68.44 },
+          { year: 2005, value: 938702, rate: 66.04 },
+          { year: 2010, value: 875062, rate: 62.47 },
+          { year: 2015, value: 796552, rate: 58.38 },
+          { year: 2020, value: 737675, rate: 55.7 },
+          { year: 2025, value: 702824, rate: 55.58 },
+          { year: 2030, value: 654137, rate: 54.4 },
+          { year: 2035, value: 598276, rate: 52.68 },
+          { year: 2040, value: 531278, rate: 49.83 },
+          { year: 2045, value: 483398, rate: 48.43 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 52449, rate: 6.72 },
+          { year: 1965, value: 60604, rate: 7.34 },
+          { year: 1970, value: 74150, rate: 7.97 },
+          { year: 1975, value: 92033, rate: 8.54 },
+          { year: 1980, value: 112753, rate: 9.32 },
+          { year: 1985, value: 132109, rate: 10.12 },
+          { year: 1990, value: 159254, rate: 11.58 },
+          { year: 1995, value: 198192, rate: 13.85 },
+          { year: 2000, value: 239432, rate: 16.6 },
+          { year: 2005, value: 283528, rate: 19.95 },
+          { year: 2010, value: 333746, rate: 23.83 },
+          { year: 2015, value: 388614, rate: 28.48 },
+          { year: 2020, value: 414576, rate: 31.3 },
+          { year: 2025, value: 420587, rate: 33.26 },
+          { year: 2030, value: 419512, rate: 34.89 },
+          { year: 2035, value: 418694, rate: 36.87 },
+          { year: 2040, value: 423548, rate: 39.72 },
+          { year: 2045, value: 410363, rate: 41.12 },
+        ],
+      },
+    ],
+  },
+
+  30: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1002191 },
+          { year: 1965, value: 1026975 },
+          { year: 1970, value: 1042736 },
+          { year: 1975, value: 1072118 },
+          { year: 1980, value: 1087012 },
+          { year: 1985, value: 1087206 },
+          { year: 1990, value: 1074325 },
+          { year: 1995, value: 1080435 },
+          { year: 2000, value: 1069912 },
+          { year: 2005, value: 1035969 },
+          { year: 2010, value: 1002198 },
+          { year: 2015, value: 963579 },
+          { year: 2020, value: 922584 },
+          { year: 2025, value: 875553 },
+          { year: 2030, value: 829087 },
+          { year: 2035, value: 781816 },
+          { year: 2040, value: 734325 },
+          { year: 2045, value: 688031 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 284394, rate: 28.38 },
+          { year: 1965, value: 252165, rate: 24.55 },
+          { year: 1970, value: 243485, rate: 23.35 },
+          { year: 1975, value: 251618, rate: 23.47 },
+          { year: 1980, value: 243977, rate: 22.44 },
+          { year: 1985, value: 225208, rate: 20.71 },
+          { year: 1990, value: 192839, rate: 17.95 },
+          { year: 1995, value: 175660, rate: 16.26 },
+          { year: 2000, value: 159496, rate: 14.91 },
+          { year: 2005, value: 142670, rate: 13.77 },
+          { year: 2010, value: 128005, rate: 12.77 },
+          { year: 2015, value: 116412, rate: 12.08 },
+          { year: 2020, value: 105360, rate: 11.42 },
+          { year: 2025, value: 98126, rate: 11.21 },
+          { year: 2030, value: 89905, rate: 10.84 },
+          { year: 2035, value: 83036, rate: 10.62 },
+          { year: 2040, value: 78184, rate: 10.65 },
+          { year: 2045, value: 73113, rate: 10.63 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 645003, rate: 64.36 },
+          { year: 1965, value: 692317, rate: 67.41 },
+          { year: 1970, value: 702967, rate: 67.42 },
+          { year: 1975, value: 708817, rate: 66.11 },
+          { year: 1980, value: 714680, rate: 65.75 },
+          { year: 1985, value: 718380, rate: 66.08 },
+          { year: 1990, value: 716161, rate: 66.66 },
+          { year: 1995, value: 709092, rate: 65.63 },
+          { year: 2000, value: 683805, rate: 63.91 },
+          { year: 2005, value: 642428, rate: 62.01 },
+          { year: 2010, value: 594573, rate: 59.33 },
+          { year: 2015, value: 546279, rate: 56.69 },
+          { year: 2020, value: 503704, rate: 54.6 },
+          { year: 2025, value: 477888, rate: 54.58 },
+          { year: 2030, value: 445710, rate: 53.76 },
+          { year: 2035, value: 412121, rate: 52.71 },
+          { year: 2040, value: 370203, rate: 50.41 },
+          { year: 2045, value: 340809, rate: 49.53 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 72794, rate: 7.26 },
+          { year: 1965, value: 82493, rate: 8.03 },
+          { year: 1970, value: 96284, rate: 9.23 },
+          { year: 1975, value: 111518, rate: 10.4 },
+          { year: 1980, value: 127950, rate: 11.77 },
+          { year: 1985, value: 143533, rate: 13.2 },
+          { year: 1990, value: 164552, rate: 15.32 },
+          { year: 1995, value: 195575, rate: 18.1 },
+          { year: 2000, value: 226323, rate: 21.15 },
+          { year: 2005, value: 249473, rate: 24.08 },
+          { year: 2010, value: 270846, rate: 27.03 },
+          { year: 2015, value: 296239, rate: 30.74 },
+          { year: 2020, value: 305500, rate: 33.11 },
+          { year: 2025, value: 299539, rate: 34.21 },
+          { year: 2030, value: 293472, rate: 35.4 },
+          { year: 2035, value: 286659, rate: 36.67 },
+          { year: 2040, value: 285938, rate: 38.94 },
+          { year: 2045, value: 274109, rate: 39.84 },
+        ],
+      },
+    ],
+  },
+  31: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 599135 },
+          { year: 1965, value: 579853 },
+          { year: 1970, value: 568777 },
+          { year: 1975, value: 581311 },
+          { year: 1980, value: 604221 },
+          { year: 1985, value: 616024 },
+          { year: 1990, value: 615722 },
+          { year: 1995, value: 614929 },
+          { year: 2000, value: 613289 },
+          { year: 2005, value: 607012 },
+          { year: 2010, value: 588667 },
+          { year: 2015, value: 573441 },
+          { year: 2020, value: 553407 },
+          { year: 2025, value: 536747 },
+          { year: 2030, value: 516255 },
+          { year: 2035, value: 494893 },
+          { year: 2040, value: 472156 },
+          { year: 2045, value: 448529 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 186407, rate: 31.11 },
+          { year: 1965, value: 154589, rate: 26.66 },
+          { year: 1970, value: 131725, rate: 23.16 },
+          { year: 1975, value: 128361, rate: 22.08 },
+          { year: 1980, value: 130631, rate: 21.62 },
+          { year: 1985, value: 130668, rate: 21.21 },
+          { year: 1990, value: 118201, rate: 19.2 },
+          { year: 1995, value: 105456, rate: 17.15 },
+          { year: 2000, value: 93584, rate: 15.26 },
+          { year: 2005, value: 84823, rate: 13.97 },
+          { year: 2010, value: 77951, rate: 13.24 },
+          { year: 2015, value: 73685, rate: 12.85 },
+          { year: 2020, value: 68330, rate: 12.35 },
+          { year: 2025, value: 65340, rate: 12.17 },
+          { year: 2030, value: 61206, rate: 11.86 },
+          { year: 2035, value: 57298, rate: 11.58 },
+          { year: 2040, value: 54393, rate: 11.52 },
+          { year: 2045, value: 51497, rate: 11.48 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 366761, rate: 61.22 },
+          { year: 1965, value: 374525, rate: 64.59 },
+          { year: 1970, value: 380499, rate: 66.9 },
+          { year: 1975, value: 388155, rate: 66.77 },
+          { year: 1980, value: 398944, rate: 66.03 },
+          { year: 1985, value: 400717, rate: 65.05 },
+          { year: 1990, value: 397218, rate: 64.51 },
+          { year: 1995, value: 390964, rate: 63.58 },
+          { year: 2000, value: 383921, rate: 62.6 },
+          { year: 2005, value: 375539, rate: 61.87 },
+          { year: 2010, value: 352098, rate: 59.81 },
+          { year: 2015, value: 326301, rate: 56.9 },
+          { year: 2020, value: 300002, rate: 54.21 },
+          { year: 2025, value: 288914, rate: 53.83 },
+          { year: 2030, value: 274780, rate: 53.23 },
+          { year: 2035, value: 261336, rate: 52.81 },
+          { year: 2040, value: 240999, rate: 51.04 },
+          { year: 2045, value: 223463, rate: 49.82 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 45967, rate: 7.67 },
+          { year: 1965, value: 50739, rate: 8.75 },
+          { year: 1970, value: 56553, rate: 9.94 },
+          { year: 1975, value: 64720, rate: 11.13 },
+          { year: 1980, value: 74474, rate: 12.33 },
+          { year: 1985, value: 84609, rate: 13.73 },
+          { year: 1990, value: 99728, rate: 16.2 },
+          { year: 1995, value: 118380, rate: 19.25 },
+          { year: 2000, value: 134984, rate: 22.01 },
+          { year: 2005, value: 146113, rate: 24.07 },
+          { year: 2010, value: 153614, rate: 26.1 },
+          { year: 2015, value: 169092, rate: 29.49 },
+          { year: 2020, value: 177046, rate: 31.99 },
+          { year: 2025, value: 182493, rate: 34 },
+          { year: 2030, value: 180269, rate: 34.92 },
+          { year: 2035, value: 176259, rate: 35.62 },
+          { year: 2040, value: 176764, rate: 37.44 },
+          { year: 2045, value: 173569, rate: 38.7 },
+        ],
+      },
+    ],
+  },
+  32: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 888886 },
+          { year: 1965, value: 821620 },
+          { year: 1970, value: 773575 },
+          { year: 1975, value: 768886 },
+          { year: 1980, value: 784795 },
+          { year: 1985, value: 794629 },
+          { year: 1990, value: 781021 },
+          { year: 1995, value: 771441 },
+          { year: 2000, value: 761503 },
+          { year: 2005, value: 742223 },
+          { year: 2010, value: 717397 },
+          { year: 2015, value: 694352 },
+          { year: 2020, value: 671126 },
+          { year: 2025, value: 642787 },
+          { year: 2030, value: 615424 },
+          { year: 2035, value: 587556 },
+          { year: 2040, value: 558290 },
+          { year: 2045, value: 528988 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 282596, rate: 31.79 },
+          { year: 1965, value: 218403, rate: 26.58 },
+          { year: 1970, value: 178457, rate: 23.07 },
+          { year: 1975, value: 168072, rate: 21.86 },
+          { year: 1980, value: 167310, rate: 21.32 },
+          { year: 1985, value: 162817, rate: 20.49 },
+          { year: 1990, value: 143884, rate: 18.42 },
+          { year: 1995, value: 126403, rate: 16.39 },
+          { year: 2000, value: 111982, rate: 14.71 },
+          { year: 2005, value: 100542, rate: 13.55 },
+          { year: 2010, value: 92218, rate: 12.85 },
+          { year: 2015, value: 86056, rate: 12.39 },
+          { year: 2020, value: 81641, rate: 12.16 },
+          { year: 2025, value: 76203, rate: 11.86 },
+          { year: 2030, value: 71080, rate: 11.55 },
+          { year: 2035, value: 66336, rate: 11.29 },
+          { year: 2040, value: 62832, rate: 11.25 },
+          { year: 2045, value: 59392, rate: 11.23 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 531573, rate: 59.8 },
+          { year: 1965, value: 523286, rate: 63.69 },
+          { year: 1970, value: 508173, rate: 65.69 },
+          { year: 1975, value: 504941, rate: 65.67 },
+          { year: 1980, value: 509938, rate: 64.98 },
+          { year: 1985, value: 510054, rate: 64.19 },
+          { year: 1990, value: 494253, rate: 63.28 },
+          { year: 1995, value: 477919, rate: 61.95 },
+          { year: 2000, value: 460103, rate: 60.42 },
+          { year: 2005, value: 439471, rate: 59.21 },
+          { year: 2010, value: 414153, rate: 57.73 },
+          { year: 2015, value: 376877, rate: 54.28 },
+          { year: 2020, value: 352656, rate: 52.55 },
+          { year: 2025, value: 335195, rate: 52.15 },
+          { year: 2030, value: 319377, rate: 51.9 },
+          { year: 2035, value: 303960, rate: 51.73 },
+          { year: 2040, value: 280285, rate: 50.2 },
+          { year: 2045, value: 260458, rate: 49.24 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 74717, rate: 8.41 },
+          { year: 1965, value: 79931, rate: 9.73 },
+          { year: 1970, value: 86945, rate: 11.24 },
+          { year: 1975, value: 95831, rate: 12.46 },
+          { year: 1980, value: 107479, rate: 13.7 },
+          { year: 1985, value: 121744, rate: 15.32 },
+          { year: 1990, value: 142061, rate: 18.19 },
+          { year: 1995, value: 167040, rate: 21.65 },
+          { year: 2000, value: 189031, rate: 24.82 },
+          { year: 2005, value: 201103, rate: 27.09 },
+          { year: 2010, value: 207398, rate: 28.91 },
+          { year: 2015, value: 222648, rate: 32.07 },
+          { year: 2020, value: 227881, rate: 33.96 },
+          { year: 2025, value: 231389, rate: 36 },
+          { year: 2030, value: 224967, rate: 36.55 },
+          { year: 2035, value: 217260, rate: 36.98 },
+          { year: 2040, value: 215173, rate: 38.54 },
+          { year: 2045, value: 209138, rate: 39.54 },
+        ],
+      },
+    ],
+  },
+  33: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1670454 },
+          { year: 1965, value: 1645135 },
+          { year: 1970, value: 1707026 },
+          { year: 1975, value: 1814305 },
+          { year: 1980, value: 1871023 },
+          { year: 1985, value: 1916906 },
+          { year: 1990, value: 1925877 },
+          { year: 1995, value: 1950750 },
+          { year: 2000, value: 1950828 },
+          { year: 2005, value: 1957264 },
+          { year: 2010, value: 1945276 },
+          { year: 2015, value: 1921525 },
+          { year: 2020, value: 1888432 },
+          { year: 2025, value: 1846230 },
+          { year: 2030, value: 1796872 },
+          { year: 2035, value: 1741517 },
+          { year: 2040, value: 1681383 },
+          { year: 2045, value: 1620031 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 480141, rate: 28.74 },
+          { year: 1965, value: 396579, rate: 24.11 },
+          { year: 1970, value: 384967, rate: 22.55 },
+          { year: 1975, value: 416900, rate: 22.98 },
+          { year: 1980, value: 424254, rate: 22.67 },
+          { year: 1985, value: 409216, rate: 21.35 },
+          { year: 1990, value: 353191, rate: 18.34 },
+          { year: 1995, value: 315902, rate: 16.19 },
+          { year: 2000, value: 291346, rate: 14.93 },
+          { year: 2005, value: 275743, rate: 14.09 },
+          { year: 2010, value: 264853, rate: 13.62 },
+          { year: 2015, value: 247890, rate: 12.9 },
+          { year: 2020, value: 229352, rate: 12.15 },
+          { year: 2025, value: 222910, rate: 12.07 },
+          { year: 2030, value: 211784, rate: 11.79 },
+          { year: 2035, value: 201661, rate: 11.58 },
+          { year: 2040, value: 195127, rate: 11.61 },
+          { year: 2045, value: 187670, rate: 11.58 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1064304, rate: 63.71 },
+          { year: 1965, value: 1105115, rate: 67.17 },
+          { year: 1970, value: 1155806, rate: 67.71 },
+          { year: 1975, value: 1203853, rate: 66.35 },
+          { year: 1980, value: 1223235, rate: 65.38 },
+          { year: 1985, value: 1257858, rate: 65.62 },
+          { year: 1990, value: 1286221, rate: 66.79 },
+          { year: 1995, value: 1294239, rate: 66.35 },
+          { year: 2000, value: 1265122, rate: 64.85 },
+          { year: 2005, value: 1236318, rate: 63.17 },
+          { year: 2010, value: 1178493, rate: 60.58 },
+          { year: 2015, value: 1098140, rate: 57.15 },
+          { year: 2020, value: 1032394, rate: 54.67 },
+          { year: 2025, value: 1044634, rate: 56.58 },
+          { year: 2030, value: 1012442, rate: 56.34 },
+          { year: 2035, value: 970500, rate: 55.73 },
+          { year: 2040, value: 899617, rate: 53.5 },
+          { year: 2045, value: 848928, rate: 52.4 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 126009, rate: 7.54 },
+          { year: 1965, value: 143441, rate: 8.72 },
+          { year: 1970, value: 166253, rate: 9.74 },
+          { year: 1975, value: 193289, rate: 10.65 },
+          { year: 1980, value: 223306, rate: 11.93 },
+          { year: 1985, value: 249555, rate: 13.02 },
+          { year: 1990, value: 285764, rate: 14.84 },
+          { year: 1995, value: 339313, rate: 17.39 },
+          { year: 2000, value: 393658, rate: 20.18 },
+          { year: 2005, value: 438054, rate: 22.38 },
+          { year: 2010, value: 484718, rate: 24.92 },
+          { year: 2015, value: 540876, rate: 28.15 },
+          { year: 2020, value: 557991, rate: 29.55 },
+          { year: 2025, value: 578686, rate: 31.34 },
+          { year: 2030, value: 572646, rate: 31.87 },
+          { year: 2035, value: 569356, rate: 32.69 },
+          { year: 2040, value: 586639, rate: 34.89 },
+          { year: 2045, value: 583433, rate: 36.01 },
+        ],
+      },
+    ],
+  },
+  34: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 2184043 },
+          { year: 1965, value: 2281146 },
+          { year: 1970, value: 2436135 },
+          { year: 1975, value: 2646324 },
+          { year: 1980, value: 2739161 },
+          { year: 1985, value: 2819200 },
+          { year: 1990, value: 2849847 },
+          { year: 1995, value: 2881748 },
+          { year: 2000, value: 2878915 },
+          { year: 2005, value: 2876642 },
+          { year: 2010, value: 2860750 },
+          { year: 2015, value: 2843990 },
+          { year: 2020, value: 2799702 },
+          { year: 2025, value: 2758220 },
+          { year: 2030, value: 2689284 },
+          { year: 2035, value: 2609116 },
+          { year: 2040, value: 2520520 },
+          { year: 2045, value: 2428818 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 630494, rate: 28.87 },
+          { year: 1965, value: 552521, rate: 24.22 },
+          { year: 1970, value: 558818, rate: 22.94 },
+          { year: 1975, value: 633219, rate: 23.93 },
+          { year: 1980, value: 647154, rate: 23.63 },
+          { year: 1985, value: 615159, rate: 21.82 },
+          { year: 1990, value: 525256, rate: 18.43 },
+          { year: 1995, value: 466553, rate: 16.19 },
+          { year: 2000, value: 428035, rate: 14.87 },
+          { year: 2005, value: 403271, rate: 14.02 },
+          { year: 2010, value: 386810, rate: 13.52 },
+          { year: 2015, value: 375890, rate: 13.22 },
+          { year: 2020, value: 352678, rate: 12.6 },
+          { year: 2025, value: 344009, rate: 12.47 },
+          { year: 2030, value: 325765, rate: 12.11 },
+          { year: 2035, value: 310624, rate: 11.91 },
+          { year: 2040, value: 301374, rate: 11.96 },
+          { year: 2045, value: 290668, rate: 11.97 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1398637, rate: 64.04 },
+          { year: 1965, value: 1552711, rate: 68.07 },
+          { year: 1970, value: 1676536, rate: 68.82 },
+          { year: 1975, value: 1777306, rate: 67.16 },
+          { year: 1980, value: 1811865, rate: 66.15 },
+          { year: 1985, value: 1879843, rate: 66.68 },
+          { year: 1990, value: 1936818, rate: 67.96 },
+          { year: 1995, value: 1956268, rate: 67.88 },
+          { year: 2000, value: 1916796, rate: 66.58 },
+          { year: 2005, value: 1858849, rate: 64.62 },
+          { year: 2010, value: 1765036, rate: 61.7 },
+          { year: 2015, value: 1662522, rate: 58.46 },
+          { year: 2020, value: 1580054, rate: 56.44 },
+          { year: 2025, value: 1578899, rate: 57.24 },
+          { year: 2030, value: 1532933, rate: 57 },
+          { year: 2035, value: 1466304, rate: 56.2 },
+          { year: 2040, value: 1359166, rate: 53.92 },
+          { year: 2045, value: 1283476, rate: 52.84 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 154912, rate: 7.09 },
+          { year: 1965, value: 175914, rate: 7.71 },
+          { year: 1970, value: 200781, rate: 8.24 },
+          { year: 1975, value: 235015, rate: 8.88 },
+          { year: 1980, value: 279132, rate: 10.19 },
+          { year: 1985, value: 323455, rate: 11.47 },
+          { year: 1990, value: 381477, rate: 13.39 },
+          { year: 1995, value: 456497, rate: 15.84 },
+          { year: 2000, value: 531537, rate: 18.46 },
+          { year: 2005, value: 600545, rate: 20.88 },
+          { year: 2010, value: 676660, rate: 23.65 },
+          { year: 2015, value: 774440, rate: 27.23 },
+          { year: 2020, value: 811931, rate: 29 },
+          { year: 2025, value: 835312, rate: 30.28 },
+          { year: 2030, value: 830586, rate: 30.89 },
+          { year: 2035, value: 832188, rate: 31.9 },
+          { year: 2040, value: 859980, rate: 34.12 },
+          { year: 2045, value: 854674, rate: 35.19 },
+        ],
+      },
+    ],
+  },
+  35: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1602207 },
+          { year: 1965, value: 1543573 },
+          { year: 1970, value: 1511448 },
+          { year: 1975, value: 1555218 },
+          { year: 1980, value: 1587079 },
+          { year: 1985, value: 1601627 },
+          { year: 1990, value: 1572616 },
+          { year: 1995, value: 1555543 },
+          { year: 2000, value: 1527964 },
+          { year: 2005, value: 1492606 },
+          { year: 2010, value: 1451338 },
+          { year: 2015, value: 1404729 },
+          { year: 2020, value: 1342059 },
+          { year: 2025, value: 1292809 },
+          { year: 2030, value: 1230324 },
+          { year: 2035, value: 1165692 },
+          { year: 2040, value: 1099619 },
+          { year: 2045, value: 1035661 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 490455, rate: 30.61 },
+          { year: 1965, value: 393960, rate: 25.52 },
+          { year: 1970, value: 348362, rate: 23.05 },
+          { year: 1975, value: 355657, rate: 22.87 },
+          { year: 1980, value: 354404, rate: 22.33 },
+          { year: 1985, value: 333923, rate: 20.85 },
+          { year: 1990, value: 278562, rate: 17.71 },
+          { year: 1995, value: 240469, rate: 15.46 },
+          { year: 2000, value: 213578, rate: 13.98 },
+          { year: 2005, value: 196729, rate: 13.18 },
+          { year: 2010, value: 184049, rate: 12.68 },
+          { year: 2015, value: 170022, rate: 12.1 },
+          { year: 2020, value: 153608, rate: 11.45 },
+          { year: 2025, value: 144656, rate: 11.19 },
+          { year: 2030, value: 133287, rate: 10.83 },
+          { year: 2035, value: 123190, rate: 10.57 },
+          { year: 2040, value: 116025, rate: 10.55 },
+          { year: 2045, value: 109151, rate: 10.54 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1002957, rate: 62.6 },
+          { year: 1965, value: 1027984, rate: 66.6 },
+          { year: 1970, value: 1025839, rate: 67.87 },
+          { year: 1975, value: 1041291, rate: 66.95 },
+          { year: 1980, value: 1048137, rate: 66.04 },
+          { year: 1985, value: 1055436, rate: 65.9 },
+          { year: 1990, value: 1042910, rate: 66.32 },
+          { year: 1995, value: 1018839, rate: 65.5 },
+          { year: 2000, value: 974131, rate: 63.75 },
+          { year: 2005, value: 920531, rate: 61.67 },
+          { year: 2010, value: 857956, rate: 59.11 },
+          { year: 2015, value: 778430, rate: 55.41 },
+          { year: 2020, value: 709992, rate: 52.9 },
+          { year: 2025, value: 688805, rate: 53.28 },
+          { year: 2030, value: 655033, rate: 53.24 },
+          { year: 2035, value: 615659, rate: 52.81 },
+          { year: 2040, value: 558879, rate: 50.82 },
+          { year: 2045, value: 515481, rate: 49.77 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 108795, rate: 6.79 },
+          { year: 1965, value: 121629, rate: 7.88 },
+          { year: 1970, value: 137247, rate: 9.08 },
+          { year: 1975, value: 158042, rate: 10.16 },
+          { year: 1980, value: 184160, rate: 11.6 },
+          { year: 1985, value: 212237, rate: 13.25 },
+          { year: 1990, value: 249488, rate: 15.86 },
+          { year: 1995, value: 295702, rate: 19.01 },
+          { year: 2000, value: 339836, rate: 22.24 },
+          { year: 2005, value: 373346, rate: 25.01 },
+          { year: 2010, value: 404694, rate: 27.88 },
+          { year: 2015, value: 447862, rate: 31.88 },
+          { year: 2020, value: 459957, rate: 34.27 },
+          { year: 2025, value: 459348, rate: 35.53 },
+          { year: 2030, value: 442004, rate: 35.93 },
+          { year: 2035, value: 426843, rate: 36.62 },
+          { year: 2040, value: 424715, rate: 38.62 },
+          { year: 2045, value: 411029, rate: 39.69 },
+        ],
+      },
+    ],
+  },
+  36: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 847274 },
+          { year: 1965, value: 815115 },
+          { year: 1970, value: 791111 },
+          { year: 1975, value: 805166 },
+          { year: 1980, value: 825261 },
+          { year: 1985, value: 834889 },
+          { year: 1990, value: 831598 },
+          { year: 1995, value: 832427 },
+          { year: 2000, value: 824108 },
+          { year: 2005, value: 809950 },
+          { year: 2010, value: 785491 },
+          { year: 2015, value: 755733 },
+          { year: 2020, value: 719559 },
+          { year: 2025, value: 687680 },
+          { year: 2030, value: 651234 },
+          { year: 2035, value: 613686 },
+          { year: 2040, value: 574474 },
+          { year: 2045, value: 535370 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 275274, rate: 32.49 },
+          { year: 1965, value: 219279, rate: 26.9 },
+          { year: 1970, value: 183878, rate: 23.24 },
+          { year: 1975, value: 177555, rate: 22.05 },
+          { year: 1980, value: 175295, rate: 21.24 },
+          { year: 1985, value: 170062, rate: 20.37 },
+          { year: 1990, value: 149770, rate: 18.01 },
+          { year: 1995, value: 132495, rate: 15.92 },
+          { year: 2000, value: 117217, rate: 14.22 },
+          { year: 2005, value: 105814, rate: 13.06 },
+          { year: 2010, value: 96596, rate: 12.3 },
+          { year: 2015, value: 87030, rate: 11.52 },
+          { year: 2020, value: 77129, rate: 10.72 },
+          { year: 2025, value: 72901, rate: 10.6 },
+          { year: 2030, value: 66597, rate: 10.23 },
+          { year: 2035, value: 60805, rate: 9.91 },
+          { year: 2040, value: 56412, rate: 9.82 },
+          { year: 2045, value: 52214, rate: 9.75 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 508733, rate: 60.04 },
+          { year: 1965, value: 527617, rate: 64.73 },
+          { year: 1970, value: 531168, rate: 67.14 },
+          { year: 1975, value: 541049, rate: 67.2 },
+          { year: 1980, value: 550779, rate: 66.74 },
+          { year: 1985, value: 553858, rate: 66.34 },
+          { year: 1990, value: 551067, rate: 66.27 },
+          { year: 1995, value: 541945, rate: 65.1 },
+          { year: 2000, value: 525724, rate: 63.79 },
+          { year: 2005, value: 506642, rate: 62.55 },
+          { year: 2010, value: 471788, rate: 60.06 },
+          { year: 2015, value: 428059, rate: 56.64 },
+          { year: 2020, value: 375657, rate: 52.21 },
+          { year: 2025, value: 370243, rate: 53.84 },
+          { year: 2030, value: 345806, rate: 53.1 },
+          { year: 2035, value: 321029, rate: 52.31 },
+          { year: 2040, value: 287977, rate: 50.13 },
+          { year: 2045, value: 261041, rate: 48.76 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 63267, rate: 7.47 },
+          { year: 1965, value: 68219, rate: 8.37 },
+          { year: 1970, value: 76065, rate: 9.61 },
+          { year: 1975, value: 86505, rate: 10.74 },
+          { year: 1980, value: 98904, rate: 11.98 },
+          { year: 1985, value: 110921, rate: 13.29 },
+          { year: 1990, value: 129105, rate: 15.52 },
+          { year: 1995, value: 157461, rate: 18.92 },
+          { year: 2000, value: 180637, rate: 21.92 },
+          { year: 2005, value: 197313, rate: 24.36 },
+          { year: 2010, value: 209926, rate: 26.73 },
+          { year: 2015, value: 230914, rate: 30.55 },
+          { year: 2020, value: 238346, rate: 33.12 },
+          { year: 2025, value: 244536, rate: 35.56 },
+          { year: 2030, value: 238831, rate: 36.67 },
+          { year: 2035, value: 231852, rate: 37.78 },
+          { year: 2040, value: 230085, rate: 40.05 },
+          { year: 2045, value: 222115, rate: 41.49 },
+        ],
+      },
+    ],
+  },
+  37: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 918867 },
+          { year: 1965, value: 900845 },
+          { year: 1970, value: 907897 },
+          { year: 1975, value: 961292 },
+          { year: 1980, value: 999864 },
+          { year: 1985, value: 1022569 },
+          { year: 1990, value: 1023412 },
+          { year: 1995, value: 1027006 },
+          { year: 2000, value: 1022890 },
+          { year: 2005, value: 1012400 },
+          { year: 2010, value: 995842 },
+          { year: 2015, value: 976263 },
+          { year: 2020, value: 950244 },
+          { year: 2025, value: 921343 },
+          { year: 2030, value: 888509 },
+          { year: 2035, value: 853054 },
+          { year: 2040, value: 814677 },
+          { year: 2045, value: 776478 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 277844, rate: 30.24 },
+          { year: 1965, value: 219450, rate: 24.36 },
+          { year: 1970, value: 201068, rate: 22.15 },
+          { year: 1975, value: 213857, rate: 22.25 },
+          { year: 1980, value: 222244, rate: 22.23 },
+          { year: 1985, value: 214695, rate: 21 },
+          { year: 1990, value: 184729, rate: 18.05 },
+          { year: 1995, value: 161674, rate: 15.74 },
+          { year: 2000, value: 148215, rate: 14.49 },
+          { year: 2005, value: 139505, rate: 13.78 },
+          { year: 2010, value: 131670, rate: 13.22 },
+          { year: 2015, value: 122324, rate: 12.53 },
+          { year: 2020, value: 113757, rate: 11.97 },
+          { year: 2025, value: 106720, rate: 11.58 },
+          { year: 2030, value: 99651, rate: 11.22 },
+          { year: 2035, value: 93160, rate: 10.92 },
+          { year: 2040, value: 88734, rate: 10.89 },
+          { year: 2045, value: 84537, rate: 10.89 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 572304, rate: 62.28 },
+          { year: 1965, value: 604865, rate: 67.14 },
+          { year: 1970, value: 620169, rate: 68.31 },
+          { year: 1975, value: 645957, rate: 67.2 },
+          { year: 1980, value: 658291, rate: 65.84 },
+          { year: 1985, value: 672022, rate: 65.72 },
+          { year: 1990, value: 680493, rate: 66.49 },
+          { year: 1995, value: 678404, rate: 66.06 },
+          { year: 2000, value: 659881, rate: 64.51 },
+          { year: 2005, value: 635746, rate: 62.8 },
+          { year: 2010, value: 595451, rate: 59.79 },
+          { year: 2015, value: 547844, rate: 56.12 },
+          { year: 2020, value: 518056, rate: 54.52 },
+          { year: 2025, value: 508666, rate: 55.21 },
+          { year: 2030, value: 488377, rate: 54.97 },
+          { year: 2035, value: 463602, rate: 54.35 },
+          { year: 2040, value: 424452, rate: 52.1 },
+          { year: 2045, value: 394400, rate: 50.79 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 68719, rate: 7.48 },
+          { year: 1965, value: 76530, rate: 8.5 },
+          { year: 1970, value: 86660, rate: 9.55 },
+          { year: 1975, value: 101376, rate: 10.55 },
+          { year: 1980, value: 119031, rate: 11.9 },
+          { year: 1985, value: 135696, rate: 13.27 },
+          { year: 1990, value: 157237, rate: 15.36 },
+          { year: 1995, value: 186850, rate: 18.19 },
+          { year: 2000, value: 214242, rate: 20.94 },
+          { year: 2005, value: 235508, rate: 23.26 },
+          { year: 2010, value: 253245, rate: 25.43 },
+          { year: 2015, value: 286296, rate: 29.33 },
+          { year: 2020, value: 296533, rate: 31.21 },
+          { year: 2025, value: 305957, rate: 33.21 },
+          { year: 2030, value: 300481, rate: 33.82 },
+          { year: 2035, value: 296292, rate: 34.73 },
+          { year: 2040, value: 301491, rate: 37.01 },
+          { year: 2045, value: 297541, rate: 38.32 },
+        ],
+      },
+    ],
+  },
+  38: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1500687 },
+          { year: 1965, value: 1446384 },
+          { year: 1970, value: 1418124 },
+          { year: 1975, value: 1465215 },
+          { year: 1980, value: 1506637 },
+          { year: 1985, value: 1529983 },
+          { year: 1990, value: 1515025 },
+          { year: 1995, value: 1506700 },
+          { year: 2000, value: 1493092 },
+          { year: 2005, value: 1467815 },
+          { year: 2010, value: 1431493 },
+          { year: 2015, value: 1385262 },
+          { year: 2020, value: 1334841 },
+          { year: 2025, value: 1274128 },
+          { year: 2030, value: 1212388 },
+          { year: 2035, value: 1148113 },
+          { year: 2040, value: 1080610 },
+          { year: 2045, value: 1012995 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 491067, rate: 32.72 },
+          { year: 1965, value: 394189, rate: 27.25 },
+          { year: 1970, value: 342671, rate: 24.16 },
+          { year: 1975, value: 343882, rate: 23.47 },
+          { year: 1980, value: 341262, rate: 22.65 },
+          { year: 1985, value: 325958, rate: 21.3 },
+          { year: 1990, value: 280919, rate: 18.54 },
+          { year: 1995, value: 245563, rate: 16.3 },
+          { year: 2000, value: 219340, rate: 14.69 },
+          { year: 2005, value: 200270, rate: 13.64 },
+          { year: 2010, value: 185179, rate: 12.94 },
+          { year: 2015, value: 169110, rate: 12.21 },
+          { year: 2020, value: 153764, rate: 11.52 },
+          { year: 2025, value: 142540, rate: 11.19 },
+          { year: 2030, value: 129656, rate: 10.69 },
+          { year: 2035, value: 118043, rate: 10.28 },
+          { year: 2040, value: 109277, rate: 10.11 },
+          { year: 2045, value: 101207, rate: 9.99 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 901968, rate: 60.1 },
+          { year: 1965, value: 933371, rate: 64.53 },
+          { year: 1970, value: 942241, rate: 66.44 },
+          { year: 1975, value: 968827, rate: 66.12 },
+          { year: 1980, value: 990401, rate: 65.74 },
+          { year: 1985, value: 1005968, rate: 65.75 },
+          { year: 1990, value: 1000404, rate: 66.03 },
+          { year: 1995, value: 982400, rate: 65.2 },
+          { year: 2000, value: 953189, rate: 63.84 },
+          { year: 2005, value: 914747, rate: 62.32 },
+          { year: 2010, value: 858991, rate: 60.01 },
+          { year: 2015, value: 776111, rate: 56.03 },
+          { year: 2020, value: 711738, rate: 53.32 },
+          { year: 2025, value: 685299, rate: 53.79 },
+          { year: 2030, value: 642775, rate: 53.02 },
+          { year: 2035, value: 599283, rate: 52.2 },
+          { year: 2040, value: 539391, rate: 49.92 },
+          { year: 2045, value: 491642, rate: 48.53 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 107652, rate: 7.17 },
+          { year: 1965, value: 118824, rate: 8.22 },
+          { year: 1970, value: 133212, rate: 9.39 },
+          { year: 1975, value: 152344, rate: 10.4 },
+          { year: 1980, value: 174729, rate: 11.6 },
+          { year: 1985, value: 198044, rate: 12.94 },
+          { year: 1990, value: 232726, rate: 15.36 },
+          { year: 1995, value: 278691, rate: 18.5 },
+          { year: 2000, value: 320078, rate: 21.44 },
+          { year: 2005, value: 351990, rate: 23.98 },
+          { year: 2010, value: 378591, rate: 26.45 },
+          { year: 2015, value: 417186, rate: 30.12 },
+          { year: 2020, value: 434279, rate: 32.53 },
+          { year: 2025, value: 446289, rate: 35.03 },
+          { year: 2030, value: 439957, rate: 36.29 },
+          { year: 2035, value: 430787, rate: 37.52 },
+          { year: 2040, value: 431942, rate: 39.97 },
+          { year: 2045, value: 420146, rate: 41.48 },
+        ],
+      },
+    ],
+  },
+  39: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 854595 },
+          { year: 1965, value: 812714 },
+          { year: 1970, value: 786882 },
+          { year: 1975, value: 808397 },
+          { year: 1980, value: 831275 },
+          { year: 1985, value: 839784 },
+          { year: 1990, value: 825034 },
+          { year: 1995, value: 816704 },
+          { year: 2000, value: 813949 },
+          { year: 2005, value: 796292 },
+          { year: 2010, value: 764456 },
+          { year: 2015, value: 728276 },
+          { year: 2020, value: 691527 },
+          { year: 2025, value: 652757 },
+          { year: 2030, value: 614449 },
+          { year: 2035, value: 575728 },
+          { year: 2040, value: 536443 },
+          { year: 2045, value: 498460 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 253081, rate: 29.61 },
+          { year: 1965, value: 200055, rate: 24.62 },
+          { year: 1970, value: 171991, rate: 21.86 },
+          { year: 1975, value: 173005, rate: 21.4 },
+          { year: 1980, value: 173649, rate: 20.89 },
+          { year: 1985, value: 168580, rate: 20.07 },
+          { year: 1990, value: 144276, rate: 17.49 },
+          { year: 1995, value: 125833, rate: 15.41 },
+          { year: 2000, value: 111740, rate: 13.73 },
+          { year: 2005, value: 102421, rate: 12.86 },
+          { year: 2010, value: 92798, rate: 12.14 },
+          { year: 2015, value: 83884, rate: 11.52 },
+          { year: 2020, value: 74946, rate: 10.84 },
+          { year: 2025, value: 68775, rate: 10.54 },
+          { year: 2030, value: 62306, rate: 10.14 },
+          { year: 2035, value: 56380, rate: 9.79 },
+          { year: 2040, value: 52105, rate: 9.71 },
+          { year: 2045, value: 48198, rate: 9.67 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 528882, rate: 61.89 },
+          { year: 1965, value: 530837, rate: 65.32 },
+          { year: 1970, value: 524918, rate: 66.71 },
+          { year: 1975, value: 536129, rate: 66.32 },
+          { year: 1980, value: 547780, rate: 65.9 },
+          { year: 1985, value: 548167, rate: 65.27 },
+          { year: 1990, value: 535995, rate: 64.97 },
+          { year: 1995, value: 522208, rate: 63.94 },
+          { year: 2000, value: 509050, rate: 62.54 },
+          { year: 2005, value: 487367, rate: 61.2 },
+          { year: 2010, value: 447540, rate: 58.54 },
+          { year: 2015, value: 400605, rate: 55.01 },
+          { year: 2020, value: 362200, rate: 52.38 },
+          { year: 2025, value: 343706, rate: 52.65 },
+          { year: 2030, value: 319382, rate: 51.98 },
+          { year: 2035, value: 296169, rate: 51.44 },
+          { year: 2040, value: 263431, rate: 49.11 },
+          { year: 2045, value: 237246, rate: 47.6 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 72632, rate: 8.5 },
+          { year: 1965, value: 81822, rate: 10.07 },
+          { year: 1970, value: 89973, rate: 11.43 },
+          { year: 1975, value: 98752, rate: 12.22 },
+          { year: 1980, value: 109116, rate: 13.13 },
+          { year: 1985, value: 121759, rate: 14.5 },
+          { year: 1990, value: 141508, rate: 17.15 },
+          { year: 1995, value: 167967, rate: 20.57 },
+          { year: 2000, value: 191729, rate: 23.56 },
+          { year: 2005, value: 206375, rate: 25.92 },
+          { year: 2010, value: 218148, rate: 28.54 },
+          { year: 2015, value: 237012, rate: 32.54 },
+          { year: 2020, value: 241787, rate: 34.96 },
+          { year: 2025, value: 240276, rate: 36.81 },
+          { year: 2030, value: 232761, rate: 37.88 },
+          { year: 2035, value: 223179, rate: 38.76 },
+          { year: 2040, value: 220907, rate: 41.18 },
+          { year: 2045, value: 213016, rate: 42.73 },
+        ],
+      },
+    ],
+  },
+  40: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 4006679 },
+          { year: 1965, value: 3964611 },
+          { year: 1970, value: 4027416 },
+          { year: 1975, value: 4292963 },
+          { year: 1980, value: 4553461 },
+          { year: 1985, value: 4719259 },
+          { year: 1990, value: 4811050 },
+          { year: 1995, value: 4933393 },
+          { year: 2000, value: 5015699 },
+          { year: 2005, value: 5049908 },
+          { year: 2010, value: 5071968 },
+          { year: 2015, value: 5101556 },
+          { year: 2020, value: 5135214 },
+          { year: 2025, value: 5042774 },
+          { year: 2030, value: 4955295 },
+          { year: 2035, value: 4841878 },
+          { year: 2040, value: 4704812 },
+          { year: 2045, value: 4554486 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 1257355, rate: 31.38 },
+          { year: 1965, value: 1040391, rate: 26.24 },
+          { year: 1970, value: 943395, rate: 23.42 },
+          { year: 1975, value: 1002084, rate: 23.34 },
+          { year: 1980, value: 1049782, rate: 23.05 },
+          { year: 1985, value: 1028211, rate: 21.79 },
+          { year: 1990, value: 910356, rate: 18.92 },
+          { year: 1995, value: 815170, rate: 16.52 },
+          { year: 2000, value: 742740, rate: 14.81 },
+          { year: 2005, value: 701195, rate: 13.89 },
+          { year: 2010, value: 684124, rate: 13.49 },
+          { year: 2015, value: 676045, rate: 13.25 },
+          { year: 2020, value: 662179, rate: 12.89 },
+          { year: 2025, value: 640257, rate: 12.7 },
+          { year: 2030, value: 609232, rate: 12.29 },
+          { year: 2035, value: 581006, rate: 12 },
+          { year: 2040, value: 561627, rate: 11.94 },
+          { year: 2045, value: 540881, rate: 11.88 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 2541467, rate: 63.43 },
+          { year: 1965, value: 2678982, rate: 67.57 },
+          { year: 1970, value: 2791505, rate: 69.31 },
+          { year: 1975, value: 2933745, rate: 68.34 },
+          { year: 1980, value: 3073049, rate: 67.49 },
+          { year: 1985, value: 3190270, rate: 67.6 },
+          { year: 1990, value: 3287878, rate: 68.34 },
+          { year: 1995, value: 3382470, rate: 68.56 },
+          { year: 2000, value: 3393080, rate: 67.65 },
+          { year: 2005, value: 3326610, rate: 65.87 },
+          { year: 2010, value: 3227932, rate: 63.64 },
+          { year: 2015, value: 3057855, rate: 59.94 },
+          { year: 2020, value: 2911353, rate: 56.69 },
+          { year: 2025, value: 2910278, rate: 57.71 },
+          { year: 2030, value: 2836752, rate: 57.25 },
+          { year: 2035, value: 2730099, rate: 56.39 },
+          { year: 2040, value: 2556972, rate: 54.35 },
+          { year: 2045, value: 2412331, rate: 52.97 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 207857, rate: 5.19 },
+          { year: 1965, value: 245238, rate: 6.19 },
+          { year: 1970, value: 292516, rate: 7.26 },
+          { year: 1975, value: 354847, rate: 8.27 },
+          { year: 1980, value: 426495, rate: 9.37 },
+          { year: 1985, value: 499228, rate: 10.58 },
+          { year: 1990, value: 597869, rate: 12.43 },
+          { year: 1995, value: 728574, rate: 14.77 },
+          { year: 2000, value: 870290, rate: 17.35 },
+          { year: 2005, value: 997798, rate: 19.76 },
+          { year: 2010, value: 1123376, rate: 22.15 },
+          { year: 2015, value: 1304764, rate: 25.58 },
+          { year: 2020, value: 1395142, rate: 27.17 },
+          { year: 2025, value: 1492239, rate: 29.59 },
+          { year: 2030, value: 1509311, rate: 30.46 },
+          { year: 2035, value: 1530773, rate: 31.62 },
+          { year: 2040, value: 1586213, rate: 33.71 },
+          { year: 2045, value: 1601274, rate: 35.16 },
+        ],
+      },
+    ],
+  },
+  41: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 942874 },
+          { year: 1965, value: 871885 },
+          { year: 1970, value: 838468 },
+          { year: 1975, value: 837674 },
+          { year: 1980, value: 865574 },
+          { year: 1985, value: 880013 },
+          { year: 1990, value: 877851 },
+          { year: 1995, value: 884316 },
+          { year: 2000, value: 876654 },
+          { year: 2005, value: 866369 },
+          { year: 2010, value: 849788 },
+          { year: 2015, value: 832832 },
+          { year: 2020, value: 811442 },
+          { year: 2025, value: 784789 },
+          { year: 2030, value: 757242 },
+          { year: 2035, value: 728170 },
+          { year: 2040, value: 696815 },
+          { year: 2045, value: 663781 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 324790, rate: 34.45 },
+          { year: 1965, value: 260148, rate: 29.84 },
+          { year: 1970, value: 214672, rate: 25.6 },
+          { year: 1975, value: 202122, rate: 24.13 },
+          { year: 1980, value: 200620, rate: 23.18 },
+          { year: 1985, value: 196114, rate: 22.29 },
+          { year: 1990, value: 177614, rate: 20.23 },
+          { year: 1995, value: 160307, rate: 18.13 },
+          { year: 2000, value: 144028, rate: 16.43 },
+          { year: 2005, value: 131969, rate: 15.23 },
+          { year: 2010, value: 123447, rate: 14.53 },
+          { year: 2015, value: 116122, rate: 13.94 },
+          { year: 2020, value: 108241, rate: 13.34 },
+          { year: 2025, value: 103830, rate: 13.23 },
+          { year: 2030, value: 97683, rate: 12.9 },
+          { year: 2035, value: 91937, rate: 12.63 },
+          { year: 2040, value: 87363, rate: 12.54 },
+          { year: 2045, value: 82750, rate: 12.47 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 558230, rate: 59.21 },
+          { year: 1965, value: 543568, rate: 62.34 },
+          { year: 1970, value: 546071, rate: 65.13 },
+          { year: 1975, value: 545454, rate: 65.12 },
+          { year: 1980, value: 562529, rate: 64.99 },
+          { year: 1985, value: 569523, rate: 64.72 },
+          { year: 1990, value: 566934, rate: 64.58 },
+          { year: 1995, value: 566671, rate: 64.08 },
+          { year: 2000, value: 553351, rate: 63.12 },
+          { year: 2005, value: 537864, rate: 62.08 },
+          { year: 2010, value: 515206, rate: 60.63 },
+          { year: 2015, value: 483019, rate: 58 },
+          { year: 2020, value: 443579, rate: 54.67 },
+          { year: 2025, value: 426949, rate: 54.4 },
+          { year: 2030, value: 406369, rate: 53.66 },
+          { year: 2035, value: 386308, rate: 53.05 },
+          { year: 2040, value: 359846, rate: 51.64 },
+          { year: 2045, value: 335720, rate: 50.58 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 59854, rate: 6.35 },
+          { year: 1965, value: 68169, rate: 7.82 },
+          { year: 1970, value: 77725, rate: 9.27 },
+          { year: 1975, value: 90026, rate: 10.75 },
+          { year: 1980, value: 102377, rate: 11.83 },
+          { year: 1985, value: 114353, rate: 12.99 },
+          { year: 1990, value: 132972, rate: 15.15 },
+          { year: 1995, value: 157329, rate: 17.79 },
+          { year: 2000, value: 179132, rate: 20.43 },
+          { year: 2005, value: 196108, rate: 22.64 },
+          { year: 2010, value: 208096, rate: 24.49 },
+          { year: 2015, value: 229335, rate: 27.54 },
+          { year: 2020, value: 246061, rate: 30.32 },
+          { year: 2025, value: 254010, rate: 32.37 },
+          { year: 2030, value: 253190, rate: 33.44 },
+          { year: 2035, value: 249925, rate: 34.32 },
+          { year: 2040, value: 249606, rate: 35.82 },
+          { year: 2045, value: 245311, rate: 36.96 },
+        ],
+      },
+    ],
+  },
+  42: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1760421 },
+          { year: 1965, value: 1641245 },
+          { year: 1970, value: 1570245 },
+          { year: 1975, value: 1571912 },
+          { year: 1980, value: 1590564 },
+          { year: 1985, value: 1593968 },
+          { year: 1990, value: 1562959 },
+          { year: 1995, value: 1544934 },
+          { year: 2000, value: 1516523 },
+          { year: 2005, value: 1478632 },
+          { year: 2010, value: 1426779 },
+          { year: 2015, value: 1377187 },
+          { year: 2020, value: 1312317 },
+          { year: 2025, value: 1257939 },
+          { year: 2030, value: 1192223 },
+          { year: 2035, value: 1124291 },
+          { year: 2040, value: 1053851 },
+          { year: 2045, value: 982200 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 638850, rate: 36.29 },
+          { year: 1965, value: 527123, rate: 32.12 },
+          { year: 1970, value: 439298, rate: 27.98 },
+          { year: 1975, value: 403824, rate: 25.69 },
+          { year: 1980, value: 385200, rate: 24.22 },
+          { year: 1985, value: 361823, rate: 22.7 },
+          { year: 1990, value: 316761, rate: 20.27 },
+          { year: 1995, value: 277263, rate: 17.95 },
+          { year: 2000, value: 243046, rate: 16.03 },
+          { year: 2005, value: 215987, rate: 14.61 },
+          { year: 2010, value: 193428, rate: 13.56 },
+          { year: 2015, value: 177562, rate: 12.89 },
+          { year: 2020, value: 164303, rate: 12.52 },
+          { year: 2025, value: 152059, rate: 12.09 },
+          { year: 2030, value: 138547, rate: 11.62 },
+          { year: 2035, value: 126629, rate: 11.26 },
+          { year: 2040, value: 117196, rate: 11.12 },
+          { year: 2045, value: 107831, rate: 10.98 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1019529, rate: 57.91 },
+          { year: 1965, value: 999303, rate: 60.89 },
+          { year: 1970, value: 1002416, rate: 63.84 },
+          { year: 1975, value: 1019301, rate: 64.84 },
+          { year: 1980, value: 1035278, rate: 65.09 },
+          { year: 1985, value: 1038396, rate: 65.15 },
+          { year: 1990, value: 1016338, rate: 65.03 },
+          { year: 1995, value: 993783, rate: 64.33 },
+          { year: 2000, value: 956692, rate: 63.08 },
+          { year: 2005, value: 913224, rate: 61.76 },
+          { year: 2010, value: 857416, rate: 60.09 },
+          { year: 2015, value: 784862, rate: 56.99 },
+          { year: 2020, value: 706077, rate: 53.8 },
+          { year: 2025, value: 663485, rate: 52.74 },
+          { year: 2030, value: 617161, rate: 51.77 },
+          { year: 2035, value: 572299, rate: 50.9 },
+          { year: 2040, value: 519310, rate: 49.28 },
+          { year: 2045, value: 475497, rate: 48.41 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 102042, rate: 5.8 },
+          { year: 1965, value: 114819, rate: 7 },
+          { year: 1970, value: 128531, rate: 8.19 },
+          { year: 1975, value: 148708, rate: 9.46 },
+          { year: 1980, value: 169753, rate: 10.67 },
+          { year: 1985, value: 193605, rate: 12.15 },
+          { year: 1990, value: 228991, rate: 14.65 },
+          { year: 1995, value: 273335, rate: 17.69 },
+          { year: 2000, value: 315871, rate: 20.83 },
+          { year: 2005, value: 348820, rate: 23.59 },
+          { year: 2010, value: 369290, rate: 25.88 },
+          { year: 2015, value: 404686, rate: 29.38 },
+          { year: 2020, value: 430353, rate: 32.79 },
+          { year: 2025, value: 442395, rate: 35.17 },
+          { year: 2030, value: 436515, rate: 36.61 },
+          { year: 2035, value: 425363, rate: 37.83 },
+          { year: 2040, value: 417345, rate: 39.6 },
+          { year: 2045, value: 398872, rate: 40.61 },
+        ],
+      },
+    ],
+  },
+  43: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1856192 },
+          { year: 1965, value: 1770736 },
+          { year: 1970, value: 1700229 },
+          { year: 1975, value: 1715273 },
+          { year: 1980, value: 1790327 },
+          { year: 1985, value: 1837747 },
+          { year: 1990, value: 1840326 },
+          { year: 1995, value: 1859793 },
+          { year: 2000, value: 1859344 },
+          { year: 2005, value: 1842233 },
+          { year: 2010, value: 1817426 },
+          { year: 2015, value: 1786170 },
+          { year: 2020, value: 1738301 },
+          { year: 2025, value: 1691314 },
+          { year: 2030, value: 1635901 },
+          { year: 2035, value: 1576825 },
+          { year: 2040, value: 1511793 },
+          { year: 2045, value: 1442442 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 629548, rate: 33.92 },
+          { year: 1965, value: 521466, rate: 29.45 },
+          { year: 1970, value: 429523, rate: 25.26 },
+          { year: 1975, value: 397402, rate: 23.17 },
+          { year: 1980, value: 395027, rate: 22.06 },
+          { year: 1985, value: 389073, rate: 21.17 },
+          { year: 1990, value: 355634, rate: 19.32 },
+          { year: 1995, value: 321462, rate: 17.28 },
+          { year: 2000, value: 288654, rate: 15.52 },
+          { year: 2005, value: 264013, rate: 14.33 },
+          { year: 2010, value: 249606, rate: 13.73 },
+          { year: 2015, value: 241167, rate: 13.5 },
+          { year: 2020, value: 228366, rate: 13.14 },
+          { year: 2025, value: 221441, rate: 13.09 },
+          { year: 2030, value: 209545, rate: 12.81 },
+          { year: 2035, value: 198731, rate: 12.6 },
+          { year: 2040, value: 190155, rate: 12.58 },
+          { year: 2045, value: 181207, rate: 12.56 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1099534, rate: 59.24 },
+          { year: 1965, value: 1105606, rate: 62.44 },
+          { year: 1970, value: 1110662, rate: 65.32 },
+          { year: 1975, value: 1134900, rate: 66.16 },
+          { year: 1980, value: 1185078, rate: 66.19 },
+          { year: 1985, value: 1206658, rate: 65.66 },
+          { year: 1990, value: 1198319, rate: 65.11 },
+          { year: 1995, value: 1196479, rate: 64.33 },
+          { year: 2000, value: 1173790, rate: 63.13 },
+          { year: 2005, value: 1139125, rate: 61.83 },
+          { year: 2010, value: 1093440, rate: 60.16 },
+          { year: 2015, value: 1024400, rate: 57.35 },
+          { year: 2020, value: 944198, rate: 54.32 },
+          { year: 2025, value: 907951, rate: 53.68 },
+          { year: 2030, value: 865823, rate: 52.93 },
+          { year: 2035, value: 826719, rate: 52.43 },
+          { year: 2040, value: 773793, rate: 51.18 },
+          { year: 2045, value: 725464, rate: 50.29 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 127110, rate: 6.85 },
+          { year: 1965, value: 143664, rate: 8.11 },
+          { year: 1970, value: 160044, rate: 9.41 },
+          { year: 1975, value: 182859, rate: 10.66 },
+          { year: 1980, value: 210051, rate: 11.73 },
+          { year: 1985, value: 241855, rate: 13.16 },
+          { year: 1990, value: 284017, rate: 15.43 },
+          { year: 1995, value: 340924, rate: 18.33 },
+          { year: 2000, value: 396020, rate: 21.3 },
+          { year: 2005, value: 437244, rate: 23.73 },
+          { year: 2010, value: 463266, rate: 25.49 },
+          { year: 2015, value: 511484, rate: 28.64 },
+          { year: 2020, value: 540538, rate: 31.1 },
+          { year: 2025, value: 561922, rate: 33.22 },
+          { year: 2030, value: 560533, rate: 34.26 },
+          { year: 2035, value: 551375, rate: 34.97 },
+          { year: 2040, value: 547845, rate: 36.24 },
+          { year: 2045, value: 535771, rate: 37.14 },
+        ],
+      },
+    ],
+  },
+  44: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1239655 },
+          { year: 1965, value: 1187480 },
+          { year: 1970, value: 1155566 },
+          { year: 1975, value: 1190314 },
+          { year: 1980, value: 1228913 },
+          { year: 1985, value: 1250214 },
+          { year: 1990, value: 1236942 },
+          { year: 1995, value: 1231306 },
+          { year: 2000, value: 1221140 },
+          { year: 2005, value: 1209571 },
+          { year: 2010, value: 1196529 },
+          { year: 2015, value: 1166338 },
+          { year: 2020, value: 1123852 },
+          { year: 2025, value: 1089084 },
+          { year: 2030, value: 1044038 },
+          { year: 2035, value: 996732 },
+          { year: 2040, value: 946917 },
+          { year: 2045, value: 896653 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 405103, rate: 32.68 },
+          { year: 1965, value: 329717, rate: 27.77 },
+          { year: 1970, value: 279605, rate: 24.2 },
+          { year: 1975, value: 274612, rate: 23.07 },
+          { year: 1980, value: 276390, rate: 22.49 },
+          { year: 1985, value: 266502, rate: 21.32 },
+          { year: 1990, value: 231265, rate: 18.7 },
+          { year: 1995, value: 200909, rate: 16.32 },
+          { year: 2000, value: 179439, rate: 14.69 },
+          { year: 2005, value: 164541, rate: 13.6 },
+          { year: 2010, value: 155634, rate: 13.01 },
+          { year: 2015, value: 146413, rate: 12.55 },
+          { year: 2020, value: 135272, rate: 12.04 },
+          { year: 2025, value: 128920, rate: 11.84 },
+          { year: 2030, value: 119767, rate: 11.47 },
+          { year: 2035, value: 111412, rate: 11.18 },
+          { year: 2040, value: 105073, rate: 11.1 },
+          { year: 2045, value: 99257, rate: 11.07 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 746271, rate: 60.2 },
+          { year: 1965, value: 759450, rate: 63.95 },
+          { year: 1970, value: 766049, rate: 66.29 },
+          { year: 1975, value: 789878, rate: 66.36 },
+          { year: 1980, value: 808130, rate: 65.76 },
+          { year: 1985, value: 819891, rate: 65.58 },
+          { year: 1990, value: 812665, rate: 65.7 },
+          { year: 1995, value: 801035, rate: 65.06 },
+          { year: 2000, value: 774403, rate: 63.42 },
+          { year: 2005, value: 748872, rate: 61.91 },
+          { year: 2010, value: 717319, rate: 59.95 },
+          { year: 2015, value: 657169, rate: 56.34 },
+          { year: 2020, value: 596527, rate: 53.08 },
+          { year: 2025, value: 581021, rate: 53.35 },
+          { year: 2030, value: 552367, rate: 52.91 },
+          { year: 2035, value: 522980, rate: 52.47 },
+          { year: 2040, value: 481160, rate: 50.81 },
+          { year: 2045, value: 444584, rate: 49.58 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 88281, rate: 7.12 },
+          { year: 1965, value: 98313, rate: 8.28 },
+          { year: 1970, value: 109912, rate: 9.51 },
+          { year: 1975, value: 125669, rate: 10.56 },
+          { year: 1980, value: 144240, rate: 11.74 },
+          { year: 1985, value: 163746, rate: 13.1 },
+          { year: 1990, value: 191441, rate: 15.48 },
+          { year: 1995, value: 229076, rate: 18.6 },
+          { year: 2000, value: 265901, rate: 21.77 },
+          { year: 2005, value: 292805, rate: 24.21 },
+          { year: 2010, value: 316750, rate: 26.47 },
+          { year: 2015, value: 351745, rate: 30.16 },
+          { year: 2020, value: 369192, rate: 32.85 },
+          { year: 2025, value: 379143, rate: 34.81 },
+          { year: 2030, value: 371904, rate: 35.62 },
+          { year: 2035, value: 362340, rate: 36.35 },
+          { year: 2040, value: 360684, rate: 38.09 },
+          { year: 2045, value: 352812, rate: 39.35 },
+        ],
+      },
+    ],
+  },
+  45: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1134590 },
+          { year: 1965, value: 1080692 },
+          { year: 1970, value: 1051105 },
+          { year: 1975, value: 1085055 },
+          { year: 1980, value: 1151587 },
+          { year: 1985, value: 1175543 },
+          { year: 1990, value: 1168907 },
+          { year: 1995, value: 1175819 },
+          { year: 2000, value: 1170007 },
+          { year: 2005, value: 1153042 },
+          { year: 2010, value: 1135233 },
+          { year: 2015, value: 1104069 },
+          { year: 2020, value: 1069576 },
+          { year: 2025, value: 1023170 },
+          { year: 2030, value: 976626 },
+          { year: 2035, value: 928034 },
+          { year: 2040, value: 876863 },
+          { year: 2045, value: 824806 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 405950, rate: 35.78 },
+          { year: 1965, value: 335056, rate: 31 },
+          { year: 1970, value: 278242, rate: 26.47 },
+          { year: 1975, value: 265987, rate: 24.51 },
+          { year: 1980, value: 273766, rate: 23.77 },
+          { year: 1985, value: 270084, rate: 22.98 },
+          { year: 1990, value: 239738, rate: 20.51 },
+          { year: 1995, value: 210806, rate: 17.93 },
+          { year: 2000, value: 187431, rate: 16.02 },
+          { year: 2005, value: 169075, rate: 14.66 },
+          { year: 2010, value: 158588, rate: 13.97 },
+          { year: 2015, value: 149608, rate: 13.55 },
+          { year: 2020, value: 139773, rate: 13.07 },
+          { year: 2025, value: 132392, rate: 12.94 },
+          { year: 2030, value: 122197, rate: 12.51 },
+          { year: 2035, value: 112605, rate: 12.13 },
+          { year: 2040, value: 104925, rate: 11.97 },
+          { year: 2045, value: 97973, rate: 11.88 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 661023, rate: 58.26 },
+          { year: 1965, value: 668656, rate: 61.87 },
+          { year: 1970, value: 684091, rate: 65.08 },
+          { year: 1975, value: 716077, rate: 65.99 },
+          { year: 1980, value: 757074, rate: 65.74 },
+          { year: 1985, value: 764617, rate: 65.04 },
+          { year: 1990, value: 761368, rate: 65.14 },
+          { year: 1995, value: 760751, rate: 64.7 },
+          { year: 2000, value: 740401, rate: 63.28 },
+          { year: 2005, value: 712527, rate: 61.8 },
+          { year: 2010, value: 680854, rate: 59.97 },
+          { year: 2015, value: 622544, rate: 56.39 },
+          { year: 2020, value: 568255, rate: 53.13 },
+          { year: 2025, value: 532963, rate: 52.09 },
+          { year: 2030, value: 500317, rate: 51.23 },
+          { year: 2035, value: 471452, rate: 50.8 },
+          { year: 2040, value: 432511, rate: 49.32 },
+          { year: 2045, value: 397299, rate: 48.17 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 67617, rate: 5.96 },
+          { year: 1965, value: 76980, rate: 7.12 },
+          { year: 1970, value: 88772, rate: 8.45 },
+          { year: 1975, value: 102923, rate: 9.49 },
+          { year: 1980, value: 120638, rate: 10.48 },
+          { year: 1985, value: 140747, rate: 11.97 },
+          { year: 1990, value: 166539, rate: 14.25 },
+          { year: 1995, value: 204222, rate: 17.37 },
+          { year: 2000, value: 241754, rate: 20.66 },
+          { year: 2005, value: 270586, rate: 23.47 },
+          { year: 2010, value: 291301, rate: 25.66 },
+          { year: 2015, value: 322975, rate: 29.25 },
+          { year: 2020, value: 344543, rate: 32.21 },
+          { year: 2025, value: 357815, rate: 34.97 },
+          { year: 2030, value: 354112, rate: 36.26 },
+          { year: 2035, value: 343977, rate: 37.07 },
+          { year: 2040, value: 339427, rate: 38.71 },
+          { year: 2045, value: 329534, rate: 39.95 },
+        ],
+      },
+    ],
+  },
+  46: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 1963104 },
+          { year: 1965, value: 1853541 },
+          { year: 1970, value: 1729150 },
+          { year: 1975, value: 1723902 },
+          { year: 1980, value: 1784623 },
+          { year: 1985, value: 1819270 },
+          { year: 1990, value: 1797824 },
+          { year: 1995, value: 1794224 },
+          { year: 2000, value: 1786194 },
+          { year: 2005, value: 1753179 },
+          { year: 2010, value: 1706242 },
+          { year: 2015, value: 1648177 },
+          { year: 2020, value: 1588256 },
+          { year: 2025, value: 1510970 },
+          { year: 2030, value: 1436753 },
+          { year: 2035, value: 1361575 },
+          { year: 2040, value: 1284036 },
+          { year: 2045, value: 1204146 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 718263, rate: 36.59 },
+          { year: 1965, value: 594204, rate: 32.06 },
+          { year: 1970, value: 468966, rate: 27.12 },
+          { year: 1975, value: 410228, rate: 23.8 },
+          { year: 1980, value: 399511, rate: 22.39 },
+          { year: 1985, value: 393154, rate: 21.61 },
+          { year: 1990, value: 357453, rate: 19.88 },
+          { year: 1995, value: 319918, rate: 17.83 },
+          { year: 2000, value: 280717, rate: 15.72 },
+          { year: 2005, value: 252285, rate: 14.39 },
+          { year: 2010, value: 233379, rate: 13.68 },
+          { year: 2015, value: 220751, rate: 13.39 },
+          { year: 2020, value: 205381, rate: 12.93 },
+          { year: 2025, value: 195052, rate: 12.91 },
+          { year: 2030, value: 179168, rate: 12.47 },
+          { year: 2035, value: 164225, rate: 12.06 },
+          { year: 2040, value: 151622, rate: 11.81 },
+          { year: 2045, value: 139987, rate: 11.63 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 1103720, rate: 56.22 },
+          { year: 1965, value: 1102221, rate: 59.47 },
+          { year: 1970, value: 1085707, rate: 62.79 },
+          { year: 1975, value: 1114775, rate: 64.67 },
+          { year: 1980, value: 1158216, rate: 64.9 },
+          { year: 1985, value: 1168410, rate: 64.22 },
+          { year: 1990, value: 1139774, rate: 63.4 },
+          { year: 1995, value: 1120432, rate: 62.45 },
+          { year: 2000, value: 1101401, rate: 61.66 },
+          { year: 2005, value: 1065960, rate: 60.8 },
+          { year: 2010, value: 1016150, rate: 59.55 },
+          { year: 2015, value: 929758, rate: 56.41 },
+          { year: 2020, value: 832194, rate: 52.4 },
+          { year: 2025, value: 784686, rate: 51.93 },
+          { year: 2030, value: 730481, rate: 50.84 },
+          { year: 2035, value: 682993, rate: 50.16 },
+          { year: 2040, value: 626685, rate: 48.81 },
+          { year: 2045, value: 572976, rate: 47.58 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 141121, rate: 7.19 },
+          { year: 1965, value: 157116, rate: 8.48 },
+          { year: 1970, value: 174477, rate: 10.09 },
+          { year: 1975, value: 198770, rate: 11.53 },
+          { year: 1980, value: 226853, rate: 12.71 },
+          { year: 1985, value: 257638, rate: 14.16 },
+          { year: 1990, value: 298904, rate: 16.63 },
+          { year: 1995, value: 353857, rate: 19.72 },
+          { year: 2000, value: 403239, rate: 22.58 },
+          { year: 2005, value: 434559, rate: 24.79 },
+          { year: 2010, value: 449692, rate: 26.36 },
+          { year: 2015, value: 479734, rate: 29.11 },
+          { year: 2020, value: 505891, rate: 31.85 },
+          { year: 2025, value: 531232, rate: 35.16 },
+          { year: 2030, value: 527104, rate: 36.69 },
+          { year: 2035, value: 514357, rate: 37.78 },
+          { year: 2040, value: 505729, rate: 39.39 },
+          { year: 2045, value: 491183, rate: 40.79 },
+        ],
+      },
+    ],
+  },
+  47: {
+    boundaryYear: 2020,
+    data: [
+      {
+        label: '総人口',
+        data: [
+          { year: 1960, value: 883122 },
+          { year: 1965, value: 934176 },
+          { year: 1970, value: 945111 },
+          { year: 1975, value: 1042572 },
+          { year: 1980, value: 1106559 },
+          { year: 1985, value: 1179097 },
+          { year: 1990, value: 1222398 },
+          { year: 1995, value: 1273440 },
+          { year: 2000, value: 1318220 },
+          { year: 2005, value: 1361594 },
+          { year: 2010, value: 1392818 },
+          { year: 2015, value: 1433566 },
+          { year: 2020, value: 1467480 },
+          { year: 2025, value: 1468236 },
+          { year: 2030, value: 1469847 },
+          { year: 2035, value: 1465761 },
+          { year: 2040, value: 1452321 },
+          { year: 2045, value: 1428305 },
+        ],
+      },
+      {
+        label: '年少人口',
+        data: [
+          { year: 1960, value: 367553, rate: 41.62 },
+          { year: 1965, value: 363048, rate: 38.86 },
+          { year: 1970, value: 329322, rate: 34.84 },
+          { year: 1975, value: 326538, rate: 31.32 },
+          { year: 1980, value: 325393, rate: 29.41 },
+          { year: 1985, value: 322523, rate: 27.35 },
+          { year: 1990, value: 299836, rate: 24.53 },
+          { year: 1995, value: 281302, rate: 22.09 },
+          { year: 2000, value: 264279, rate: 20.05 },
+          { year: 2005, value: 254203, rate: 18.67 },
+          { year: 2010, value: 246313, rate: 17.68 },
+          { year: 2015, value: 247206, rate: 17.24 },
+          { year: 2020, value: 243246, rate: 16.58 },
+          { year: 2025, value: 242835, rate: 16.54 },
+          { year: 2030, value: 235018, rate: 15.99 },
+          { year: 2035, value: 228171, rate: 15.57 },
+          { year: 2040, value: 223798, rate: 15.41 },
+          { year: 2045, value: 218111, rate: 15.27 },
+        ],
+      },
+      {
+        label: '生産年齢人口',
+        data: [
+          { year: 1960, value: 467269, rate: 52.91 },
+          { year: 1965, value: 516288, rate: 55.27 },
+          { year: 1970, value: 553486, rate: 58.56 },
+          { year: 1975, value: 642115, rate: 61.59 },
+          { year: 1980, value: 694174, rate: 62.73 },
+          { year: 1985, value: 754119, rate: 63.96 },
+          { year: 1990, value: 793287, rate: 64.9 },
+          { year: 1995, value: 842752, rate: 66.18 },
+          { year: 2000, value: 861826, rate: 65.38 },
+          { year: 2005, value: 888046, rate: 65.22 },
+          { year: 2010, value: 897960, rate: 64.47 },
+          { year: 2015, value: 892109, rate: 62.23 },
+          { year: 2020, value: 871154, rate: 59.36 },
+          { year: 2025, value: 863752, rate: 58.83 },
+          { year: 2030, value: 850472, rate: 57.86 },
+          { year: 2035, value: 830504, rate: 56.66 },
+          { year: 2040, value: 792187, rate: 54.55 },
+          { year: 2045, value: 761952, rate: 53.35 },
+        ],
+      },
+      {
+        label: '老年人口',
+        data: [
+          { year: 1960, value: 48171, rate: 5.45 },
+          { year: 1965, value: 54789, rate: 5.86 },
+          { year: 1970, value: 62303, rate: 6.59 },
+          { year: 1975, value: 72539, rate: 6.96 },
+          { year: 1980, value: 85819, rate: 7.76 },
+          { year: 1985, value: 101947, rate: 8.65 },
+          { year: 1990, value: 121082, rate: 9.91 },
+          { year: 1995, value: 148567, rate: 11.67 },
+          { year: 2000, value: 182557, rate: 13.85 },
+          { year: 2005, value: 218897, rate: 16.08 },
+          { year: 2010, value: 240507, rate: 17.27 },
+          { year: 2015, value: 278337, rate: 19.42 },
+          { year: 2020, value: 324708, rate: 22.13 },
+          { year: 2025, value: 361649, rate: 24.63 },
+          { year: 2030, value: 384357, rate: 26.15 },
+          { year: 2035, value: 407086, rate: 27.77 },
+          { year: 2040, value: 436336, rate: 30.04 },
+          { year: 2045, value: 448242, rate: 31.38 },
+        ],
+      },
+    ],
+  },
+};
+
+export const populationCompositionPerYearHandlers = [
+  http.get('*/api/resas/population/composition/perYear', async ({ request }) => {
+    const url = new URL(request.url);
+    const prefCode = url.searchParams.get('prefCode');
+    const cityCode = url.searchParams.get('cityCode');
+
+    if (!prefCode || !cityCode) {
+      return new HttpResponse(null, {
+        status: 400,
+      });
+    }
+
+    // MSWのモックサーバーでtimeoutパラメーターを指定することで、リクエストの遅延をシミュレートできます
+    const timeout = url.searchParams.get('timeout');
+    if (timeout) {
+      await delay(Number(timeout));
+    }
+
+    return HttpResponse.json({
+      message: null,
+      result: mockPopulationCompositionPerYear[Number(prefCode) as PrefCode],
+    });
+  }),
+];
