+++ src/mocks/handlers/api/test/index.ts
@@ -1,21 +1,1 @@
-import { http, HttpResponse } from 'msw';
-
-const handlers = [
-  // fetcher の挙動を確認するためのハンドラ
-  http.get('*/api/test', ({ request }) => {
-    const url = new URL(request.url);
-    const error = url.searchParams.get('error');
-
-    if (error) {
-      return new HttpResponse(null, {
-        status: 500,
-      });
-    }
-
-    return HttpResponse.json({
-      message: 'Hello, world!',
-    });
-  }),
-];
-
-export default handlers;
+export { testHandlers } from './handlers';
